+++ src/wrench/action/FileDeleteAction.cpp
@@ -27,13 +27,12 @@ namespace wrench {
     /**
     * @brief Constructor
     * @param name: the action's name (if empty, a unique name will be picked for you)
-    * @param job: the job this action belongs to
     * @param file: the file
     * @param file_location: the location where the file should be deleted
     */
-    FileDeleteAction::FileDeleteAction(const std::string &name, std::shared_ptr<CompoundJob> job,
+    FileDeleteAction::FileDeleteAction(const std::string &name,
                                        std::shared_ptr<DataFile> file,
-                                       std::shared_ptr<FileLocation> file_location) : Action(name, "file_delete_", job),
+                                       std::shared_ptr<FileLocation> file_location) : Action(name, "file_delete_"),
                                                                                       file(std::move(file)), file_location(std::move(file_location)) {
     }
 
