+++ examples/action_api/super-custom-action/SuperCustomActionController.cpp
@@ -39,9 +39,8 @@ namespace wrench {
                             unsigned long num_cores,
                             std::function<void(std::shared_ptr<ActionExecutor> action_executor)> lambda_execute,
                             std::function<void(std::shared_ptr<ActionExecutor> action_executor)> lambda_terminate,
-                            int value) :
-                CustomAction(name, ram, num_cores, lambda_execute, lambda_terminate),
-                additional_variable(value) {}
+                            int value) : CustomAction(name, ram, num_cores, lambda_execute, lambda_terminate),
+                                         additional_variable(value) {}
     };
 
     /**
@@ -154,21 +153,22 @@ namespace wrench {
 
         /* Now create a job with a derived custom action */
         auto even_more_custom_job = job_manager->createCompoundJob("even_more_custom_job");
-        auto my_derived_custom_action = std::make_shared<DerivedCustomAction>("powerful",
-                                                                              0, 0,
-                                                                              [num_cores_to_use_for_vm, cloud_service](std::shared_ptr<ActionExecutor> action_executor) {
-                                                                                  TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_RED);
+        auto my_derived_custom_action = std::make_shared<DerivedCustomAction>(
+                "powerful",
+                0, 0,
+                [num_cores_to_use_for_vm, cloud_service](std::shared_ptr<ActionExecutor> action_executor) {
+                    TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_RED);
 
-                                                                                  auto the_action = std::dynamic_pointer_cast<DerivedCustomAction>(action_executor->getAction());
-                                                                                  WRENCH_INFO("Derived custom action executing on host %s", action_executor->getHostname().c_str());
-                                                                                  WRENCH_INFO("The additional instance variable's value was %d", the_action->additional_variable);
-                                                                                  WRENCH_INFO("But I am incrementing it by 1");
-                                                                                  the_action->additional_variable++;
-                                                                              },
-                                                                              [](std::shared_ptr<ActionExecutor> action_executor) {
-                                                                                  WRENCH_INFO("Derived custom action terminating");
-                                                                              },
-                                                                              42);
+                    auto the_action = std::dynamic_pointer_cast<DerivedCustomAction>(action_executor->getAction());
+                    WRENCH_INFO("Derived custom action executing on host %s", action_executor->getHostname().c_str());
+                    WRENCH_INFO("The additional instance variable's value was %d", the_action->additional_variable);
+                    WRENCH_INFO("But I am incrementing it by 1");
+                    the_action->additional_variable++;
+                },
+                [](std::shared_ptr<ActionExecutor> action_executor) {
+                    WRENCH_INFO("Derived custom action terminating");
+                },
+                42);
         even_more_custom_job->addCustomAction(my_derived_custom_action);
 
         /* Submit the job to the bare-metal service compute service, letting the service pick where
