### 2022-04-14 7224856
Removed some CodeFactor issues
