### 2022-04-14 010895f
Removed some CodeFactor issues
