+++ include/wrench/simgrid_S4U_util/S4U_Mailbox.h
@@ -47,7 +47,7 @@ namespace wrench {
 
         static void createMailboxPool(unsigned long num_mailboxes);
 
-        static simgrid::s4u::Mailbox *generateUniqueMailbox(std::string prefix);
+        static simgrid::s4u::Mailbox *generateUniqueMailbox(const std::string& prefix);
 
         /**
          * @brief The mailbox pool size
