+++ src/wrench/execution_events/ExecutionEvent.cpp
@@ -17,7 +17,6 @@
 #include <wrench/execution_controller/ExecutionControllerMessage.h>
 #include "wrench/services/storage/StorageServiceMessage.h"
 #include <wrench/exceptions/ExecutionException.h>
-#include "wrench.h"
 
 WRENCH_LOG_CATEGORY(wrench_core_workflow_execution_event, "Log category for Workflow Execution Event");
 
