+++ src/wrench/data_file/DataFile.cpp
@@ -8,6 +8,7 @@
  */
 
 #include <map>
+#include <utility>
 #include <wrench/logging/TerminalOutput.h>
 
 #include <wrench/data_file/DataFile.h>
@@ -23,14 +24,14 @@ namespace wrench {
      * @param id: the file id
      * @param s: the file size
      */
-    DataFile::DataFile(const std::string id, double s) : id(id), size(s) {
+    DataFile::DataFile(std::string  id, double s) : id(std::move(id)), size(s) {
     }
 
     /**
      * @brief Get the file size
      * @return a size in bytes
      */
-    double DataFile::getSize() {
+    double DataFile::getSize() const {
         return this->size;
     }
 
@@ -38,7 +39,7 @@ namespace wrench {
      * @brief Get the file id
      * @return the id
      */
-    std::string DataFile::getID() {
+    std::string DataFile::getID() const {
         return this->id;
     }
 

[…]
