+++ examples/workflow_api/basic-examples/bare-metal-chain/OneTaskAtATimeWMS.cpp
@@ -47,7 +47,6 @@ namespace wrench {
      * @throw std::runtime_error
      */
     int OneTaskAtATimeWMS::main() {
-
         /* Set the logging output to GREEN */
         TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_GREEN);
 
@@ -59,7 +58,6 @@ namespace wrench {
 
         /* While the workflow isn't done, repeat the main loop */
         while (not this->workflow->isDone()) {
-
             /* Get one ready task */
             auto ready_task = this->workflow->getReadyTasks().at(0);
 
