+++ examples/workflow_api/basic-examples/virtualized-cluster-bag-of-tasks/TwoTasksAtATimeVirtualizedClusterWMS.cpp
@@ -53,7 +53,6 @@ namespace wrench {
      * @throw std::runtime_error
      */
     int TwoTasksAtATimeVirtualizedClusterWMS::main() {
-
         /* Set the logging output to GREEN */
         TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_GREEN);
 
@@ -76,7 +75,6 @@ namespace wrench {
 
         /* While the workflow isn't done, repeat the main loop */
         while (not this->workflow->isDone()) {
-
             /* Get the ready tasks */
             auto ready_tasks = this->workflow->getReadyTasks();
 
