### 2022-04-14 a2024d3
Committing clang-format changes
