+++ src/wrench/simgrid_S4U_util/S4U_Simulation.cpp
@@ -293,7 +293,7 @@ namespace wrench {
  * @param hostname: the name of the host
  * @return true or false
  */
-    bool S4U_Simulation::hostExists(const std::string& hostname) {
+    bool S4U_Simulation::hostExists(const std::string &hostname) {
         return (simgrid::s4u::Host::by_name_or_null(hostname) != nullptr);
     }
 
