+++ include/wrench/execution_controller/ExecutionController.h
@@ -74,7 +74,7 @@ namespace wrench {
     protected:
         ExecutionController(
                 const std::string &hostname,
-                const std::string& suffix);
+                const std::string &suffix);
 
 
         void setTimer(double date, std::string message);
