### 2022-04-15 bbd93c2
(#260) Implemented contingency measures in case WfCommons JSON file contains machine specs, but these machine specs don't indicate core speed!
