### 2022-04-15 3b78544
Committing clang-format changes
