### 2022-04-15 6cb9723
(#260) If no number of cores is specific in machine spec in a WfCommons JSON, just assume num_cores = 1
