+++ include/wrench/services/storage/StorageService.h
@@ -62,8 +62,7 @@ namespace wrench {
         /***********************/
         /** \cond INTERNAL    **/
         /***********************/
-
-        static void readFile(std::shared_ptr<DataFile> file, std::shared_ptr<FileLocation> location, simgrid::s4u::Mailbox * answer_mailbox, simgrid::s4u::Mailbox * chunk_receiving_mailbox, double num_bytes);
+        static void readFile(const std::shared_ptr<DataFile>& file, const std::shared_ptr<FileLocation>& location, simgrid::s4u::Mailbox * answer_mailbox, simgrid::s4u::Mailbox * chunk_receiving_mailbox, double num_bytes);
 
         bool isScratch() const;
 
