+++ src/wrench/services/storage/xrootd/SearchStack.cpp
@@ -29,6 +29,10 @@ namespace wrench {
             return *current;
 
         }
+        bool SearchStack::empty(){
+            return stack.empty();
+
+        }
         bool SearchStack::atEnd(){
             return current==stack.end();
         }
