### 2022-04-19 59bcfb0
Added a method to modify the size of a file at runtime (which is documented to be very dangerous but sometimes useful)
