### 2022-04-21 eb5561b
Added string -> message/service translation
