### 2022-04-21 d0cd0ac
Added string -> message/service translation
