+++ include/wrench/services/ServiceMessagePayload.h
@@ -12,6 +12,8 @@
 #define WRENCH_SERVICEMESSAGEPAYLOAD_H
 
 #include <string>
+#include <map>
+
 namespace wrench {
 
     /***********************/
@@ -34,7 +36,7 @@ namespace wrench {
 #define DECLARE_MESSAGEPAYLOAD_NAME(name) static const WRENCH_MESSAGEPAYLOAD_TYPE name
 
 
-#define SET_MESSAGEPAYLOAD_NAME(classname, name) const WRENCH_MESSAGEPAYLOAD_TYPE classname::name = ++WRENCH_MESSAGEPAYLOAD_COUNT
+#define SET_MESSAGEPAYLOAD_NAME(classname, name) const WRENCH_MESSAGEPAYLOAD_TYPE classname::name = classname::addMessagePayload(#classname,#name)
 //#name
 //++WRENCH_MESSAGEPAYLOAD_COUNT
 namespace wrench {
@@ -43,8 +45,11 @@ namespace wrench {
      * @brief Configurable message payloads for a Service
      */
     class ServiceMessagePayload {
-
+         static std::map<std::string,WRENCH_MESSAGEPAYLOAD_TYPE> stringToPayloadMap;
     public:
+
+        static WRENCH_MESSAGEPAYLOAD_TYPE addMessagePayload(std::string classname,std::string messagePayload);
+        static WRENCH_MESSAGEPAYLOAD_TYPE translateString(std::string messagePayload);
         /** @brief The number of bytes in the control message sent to the daemon to terminate it **/
         DECLARE_MESSAGEPAYLOAD_NAME(STOP_DAEMON_MESSAGE_PAYLOAD);
         /** @brief The number of bytes in the control message sent by the daemon to confirm it has terminated **/
