+++ src/wrench/services/ServiceMessagePayload.cpp
@@ -18,12 +18,12 @@ namespace wrench {
     std::map<std::string,WRENCH_MESSAGEPAYLOAD_TYPE> ServiceMessagePayload::stringToPayloadMap;
     SET_MESSAGEPAYLOAD_NAME(ServiceMessagePayload, STOP_DAEMON_MESSAGE_PAYLOAD);
     SET_MESSAGEPAYLOAD_NAME(ServiceMessagePayload, DAEMON_STOPPED_MESSAGE_PAYLOAD);
+
     /**
      * @brief add new message to payload map.  DO NOT CALL THIS FUNCTION DIRECTLY, use SET_MESSAGEPAYLOAD_NAME and DECLARE_MESSAGEPAYLOAD_NAME
      * @param classname: The class to add the message too
      * @param messagePayload: the name of the message payload to add
      */
-
     WRENCH_MESSAGEPAYLOAD_TYPE ServiceMessagePayload::addMessagePayload(std::string classname,std::string messagePayload){
          ++WRENCH_MESSAGEPAYLOAD_COUNT;
          stringToPayloadMap[classname+"::"+messagePayload]=WRENCH_MESSAGEPAYLOAD_COUNT;
