### 2022-04-22 40617fd
Augmented a unit test to use the wrench::ServiceMessagePayload::translateString and the wrench::ServiceProperty::translateString methods
