+++ test/simulation/SimpleSimulationTest.cpp
@@ -326,16 +326,16 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
     // Adding services to an uninitialized simulation
     std::vector<std::string> hosts = {"DualCoreHost", "QuadCoreHost"};
     ASSERT_THROW(simulation->add(
-                         new wrench::CloudComputeService("DualCoreHost", hosts, "/scratch")),
+            new wrench::CloudComputeService("DualCoreHost", hosts, "/scratch")),
                  std::runtime_error);
     ASSERT_THROW(simulation->add(
-                         new wrench::SimpleStorageService("DualCoreHost", {"/"})),
+            new wrench::SimpleStorageService("DualCoreHost", {"/"})),
                  std::runtime_error);
     ASSERT_THROW(simulation->add(
-                         new wrench::NetworkProximityService("DualCoreHost", hosts)),
+            new wrench::NetworkProximityService("DualCoreHost", hosts)),
                  std::runtime_error);
     ASSERT_THROW(simulation->add(
-                         new wrench::FileRegistryService("DualCoreHost")),
+            new wrench::FileRegistryService("DualCoreHost")),
                  std::runtime_error);
 
     ASSERT_NO_THROW(simulation->init(&argc, argv));
@@ -349,12 +349,19 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
 
     // Create a Storage Service
     ASSERT_THROW(storage_service = simulation->add(
-                         new wrench::SimpleStorageService(hostname, {"/disk1"}, {},
-                                                          {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, -1}})),
+            new wrench::SimpleStorageService(hostname, {"/disk1"}, {},
+                                             {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, -1}})),
                  std::invalid_argument);
     storage_service = simulation->add(
-            new wrench::SimpleStorageService(hostname, {"/disk2"}, {},
-                                             {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, 123}}));
+            new wrench::SimpleStorageService(hostname, {"/disk2"},
+                                             {
+                                                     {wrench::SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS, "567"},
+                                                     {wrench::ServiceProperty::translateString("wrench::SimpleStorageServiceProperty::BUFFER_SIZE"), "678"}
+                                             },
+                                             {
+                                                     {wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, 123},
+                                                     {wrench::ServiceMessagePayload::translateString("wrench::SimpleStorageServiceMessagePayload::FILE_LOOKUP_ANSWER_MESSAGE_PAYLOAD"), 234}
+                                             }));
 
 
     // Try to get a bogus property as string or double
@@ -363,16 +370,27 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
     ASSERT_THROW(storage_service->getPropertyValueAsUnsignedLong(-1), std::invalid_argument);
[…]
