### 2022-04-27 1887559
added a translatePayloadType() function to go from Payloads to string
Signed-off-by: Loïc Pottier <lpottier@isi.edu>
