+++ include/wrench/services/ServiceMessagePayload.h
@@ -46,10 +46,12 @@ namespace wrench {
      */
     class ServiceMessagePayload {
          static std::map<std::string,WRENCH_MESSAGEPAYLOAD_TYPE> stringToPayloadMap;
+         static std::map<WRENCH_MESSAGEPAYLOAD_TYPE,std::string> payloadToString;
     public:
 
         static WRENCH_MESSAGEPAYLOAD_TYPE addMessagePayload(std::string classname,std::string messagePayload);
         static WRENCH_MESSAGEPAYLOAD_TYPE translateString(std::string messagePayload);
+        static std::string translatePayloadType(WRENCH_MESSAGEPAYLOAD_TYPE messagePayload);
         /** @brief The number of bytes in the control message sent to the daemon to terminate it **/
         DECLARE_MESSAGEPAYLOAD_NAME(STOP_DAEMON_MESSAGE_PAYLOAD);
         /** @brief The number of bytes in the control message sent by the daemon to confirm it has terminated **/
