### 2022-04-27 94bdd3e
added a translatePayloadType() function to go from Payloads to string
Signed-off-by: Loïc Pottier <lpottier@isi.edu>
