### 2022-04-28 f793690
A bunch of fixes, including changing the order of files in CMakeLists.txt!! Better unit tests
