+++ test/simulation/SimpleSimulationTest.cpp
@@ -355,17 +355,17 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
             new wrench::SimpleStorageService(hostname, {"/disk1"}, {},
                                              {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, -1}})),
                  std::invalid_argument);
+
     storage_service = simulation->add(
             new wrench::SimpleStorageService(hostname, {"/disk2"},
                                              {
                                                      {wrench::SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS, "567"},
-                                                     {wrench::ServiceProperty::translateString("wrench::SimpleStorageServiceProperty::BUFFER_SIZE"), "678"}
+                                                     {wrench::ServiceProperty::translateString("StorageServiceProperty::BUFFER_SIZE"), "678"}
                                              },
                                              {
                                                      {wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, 123},
-                                                     {wrench::ServiceMessagePayload::translateString("wrench::SimpleStorageServiceMessagePayload::FILE_LOOKUP_ANSWER_MESSAGE_PAYLOAD"), 234}
+                                                     {wrench::ServiceMessagePayload::translateString("StorageServiceMessagePayload::FILE_LOOKUP_ANSWER_MESSAGE_PAYLOAD"), 234}
                                              }));
-
 
     // Try to get a bogus property as string or double
     ASSERT_THROW(storage_service->getPropertyValueAsString(-1), std::invalid_argument);
@@ -374,26 +374,50 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
     ASSERT_THROW(storage_service->getPropertyValueAsBoolean(-1), std::invalid_argument);
 
     ASSERT_EQ(storage_service->getPropertyValueAsUnsignedLong(wrench::SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS), 567);
-    ASSERT_EQ(storage_service->getPropertyValueAsUnsignedLong(wrench::ServiceProperty::translateString("wrench::SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS")), 567);
+    ASSERT_EQ(storage_service->getPropertyValueAsUnsignedLong(wrench::ServiceProperty::translateString("SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS")), 567);
     ASSERT_EQ(storage_service->getPropertyValueAsUnsignedLong(wrench::SimpleStorageServiceProperty::BUFFER_SIZE), 678);
+    ASSERT_EQ(storage_service->getPropertyValueAsUnsignedLong(wrench::ServiceProperty::translateString("StorageServiceProperty::BUFFER_SIZE")), 678);
 
     ASSERT_THROW(storage_service->getMessagePayloadValue(-1), std::invalid_argument);
     ASSERT_EQ(123, storage_service->getMessagePayloadValue(
             wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD));
     ASSERT_EQ(123, storage_service->getMessagePayloadValue(
[…]
