### 2022-04-28 340d1b9
A bunch of fixes, including changing the order of files in CMakeLists.txt!! Better unit tests
