+++ include/wrench/services/ServiceMessagePayload.h
@@ -33,10 +33,10 @@ namespace wrench {
     /***********************/
 
 }// namespace wrench
-#define DECLARE_MESSAGEPAYLOAD_NAME(name) static const WRENCH_MESSAGEPAYLOAD_TYPE name
 
+#define DECLARE_MESSAGEPAYLOAD_NAME(name) static const wrench::WRENCH_MESSAGEPAYLOAD_TYPE name
 
-#define SET_MESSAGEPAYLOAD_NAME(classname, name) const WRENCH_MESSAGEPAYLOAD_TYPE classname::name = classname::addMessagePayload(#classname,#name)
+#define SET_MESSAGEPAYLOAD_NAME(classname, name) const wrench::WRENCH_MESSAGEPAYLOAD_TYPE classname::name = classname::addMessagePayload(#classname,#name)
 //#name
 //++WRENCH_MESSAGEPAYLOAD_COUNT
 namespace wrench {
@@ -52,6 +52,7 @@ namespace wrench {
         static WRENCH_MESSAGEPAYLOAD_TYPE addMessagePayload(std::string classname,std::string messagePayload);
         static WRENCH_MESSAGEPAYLOAD_TYPE translateString(std::string messagePayload);
         static std::string translatePayloadType(WRENCH_MESSAGEPAYLOAD_TYPE messagePayload);
+
         /** @brief The number of bytes in the control message sent to the daemon to terminate it **/
         DECLARE_MESSAGEPAYLOAD_NAME(STOP_DAEMON_MESSAGE_PAYLOAD);
         /** @brief The number of bytes in the control message sent by the daemon to confirm it has terminated **/
