+++ include/wrench/workflow/Workflow.h
@@ -96,6 +96,10 @@ namespace wrench {
         std::set<std::shared_ptr<WorkflowTask>> getTasksThatInput(const std::shared_ptr<DataFile> &file);
         bool isDone();
 
+        void disableTopLevelDynamicUpdates();
+        void enableTopLevelDynamicUpdates();
+        void updateAllTopLevels();
+
         /***********************/
         /** \cond DEVELOPER    */
         /***********************/
@@ -114,6 +118,8 @@ namespace wrench {
         friend class WMS;
         friend class Simulation;
         friend class WorkflowTask;
+
+        bool update_top_levels_dynamically;
 
         Workflow();
 
