### 2022-04-30 ec30ac8
Made workflow .json file more parsable
