### 2022-05-03 076b911
segfault fix on Linux
