+++ include/wrench/services/ServiceProperty.h
@@ -31,10 +31,7 @@ namespace wrench {
 }// namespace wrench
 
 #define DECLARE_PROPERTY_NAME(name) static const wrench::WRENCH_PROPERTY_TYPE name
-//#define SET_PROPERTY_NAME(classname, name) const wrench::WRENCH_PROPERTY_TYPE classname::name = ++wrench::WRENCH_PROPERTY_COUNT
 #define SET_PROPERTY_NAME(classname, name) const wrench::WRENCH_PROPERTY_TYPE classname::name = classname::addServiceProperty(#classname, #name)
-//++wrench::WRENCH_PROPERTY_COUNT
-//#name
 
 namespace wrench {
 
@@ -44,8 +41,6 @@ namespace wrench {
 
     class ServiceProperty {
 
-        static std::map<std::string, WRENCH_PROPERTY_TYPE> stringToPropertyMap;
-        static std::map<WRENCH_PROPERTY_TYPE, std::string> propertyToStringMap;
 
     public:
         static WRENCH_PROPERTY_TYPE addServiceProperty(std::string classname, std::string messagePayload);
