+++ src/wrench/services/storage/StorageService.cpp
@@ -214,7 +214,7 @@ namespace wrench {
      * @throw std::invalid_arguments
      */
 
-    void StorageService::readFile(std::shared_ptr<DataFile> file, std::shared_ptr<FileLocation> location) {
+    void StorageService::readFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<FileLocation> &location) {
         //if ((file == nullptr) or (location == nullptr)) {
         //    throw std::invalid_argument("StorageService::readFile(): Invalid arguments");
         //}
