+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -22,7 +22,6 @@ WRENCH_LOG_CATEGORY(wrench_core_xrootd_data_server,
                     "Log category for XRootD");
 namespace wrench {
     namespace XRootD{
-//todo add specific handling for being a file server
             int Node::main() {
                 //TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_CYAN);
 
@@ -86,51 +85,7 @@ namespace wrench {
                         shared_ptr<FileLocation> best=cached[0];//use load based search, not just 0
                         StorageService::readFile(msg->file,best,msg->answer_mailbox,msg->mailbox_to_receive_the_file_content,msg->num_bytes_to_read);
                     }else{
-                        bool failed=false;
-                        auto all= metavisor->getFileNodes(msg->file);
-                        if(all.size()>0){
-                            auto subtree= searchAll(all,msg->file);
-                            if(subtree.size()>0){
-                                for(auto stack:subtree){
-                                    stack->moveDown();
-                                }
-                                auto bundles=bundle(subtree);
-                                for(auto stacks :bundles){
-                                    auto target=stacks[0]->peak();
-                                    S4U_Mailbox::putMessage(target->mailbox,new ContinueSearchMessage(msg->answer_mailbox,msg->mailbox_to_receive_the_file_content,stacks,msg->payload));
-                                }
-                            }else{
-                                failed=true;
-                                WRENCH_INFO(
-                                        "Received a read request for a file Not in XRootD subtree");
-                            }
-                            auto client_answer_mailbox = msg->answer_mailbox;
-                            auto client_mailbox_to_receive_the_file_content = msg->mailbox_to_receive_the_file_content;
 
-                            //search all
-                            //send message to all and appropriate stacks
-                        }else{
-                            failed=true;
-                            WRENCH_INFO(
-                                    "Received a read request for a file Not in XRootD tree");
-                        }
-                        if(failed){
-
-                            std::shared_ptr<FailureCause> failure_cause;
-
-                            failure_cause = std::shared_ptr<FailureCause>(new FileNotFound(msg->file, nullptr));//I dont know what to give for the location
-
-                            S4U_Mailbox::dputMessage(
-                                    msg->answer_mailbox,
-                                    new FileReadAnswerMessage(
-                                            msg->file,
-                                            nullptr,
-                                            false,
-                                            failure_cause,
-                                            this->getMessagePayloadValue(
-                                                    MessagePayload::FILE_READ_ANSWER_MESSAGE_PAYLOAD)));
-
-                        }
                     }
                     return true;
                 }  else if (auto msg = dynamic_cast<UpdateCacheMessage *>(message.get())) {
@@ -146,39 +101,14 @@ namespace wrench {
                     return true;
 
                 }  else if (auto msg = dynamic_cast<ContinueSearchMessage *>(message.get())) {
-                    //add queue checking
-//                    if(msg->path.size()>0){//continue search in "children"
-//                        auto targets=msg->path.top();
-//                        msg->path.pop();
-//                        //filter repeat sends
-//                        for(auto target:targets){
-//                            S4U_Mailbox::putMessage(target->mailbox,msg);
-//                        }
-//                    }
-//                    if(internalStorage){//check own file system
-//                        auto potentialLocaiton=FileLocation::LOCATION(internalStorage);
-//                        if(StorageService::lookupFile(msg->file,potentialLocaiton)){
-//                            if(supervisor&&this!=msg->cache_to){
[…]
