### 2022-05-12 d267512
added Naive implimentation of XRootD
