+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -18,6 +18,8 @@
 #include "wrench/services/storage/xrootd/XRootDMessage.h"
 #include "wrench/services/storage/StorageServiceMessage.h"
 #include "wrench/services/storage/xrootd/SearchStack.h"
+
+#include <wrench/exceptions/ExecutionException.h>
 WRENCH_LOG_CATEGORY(wrench_core_xrootd_data_server,
                     "Log category for XRootD");
 namespace wrench {
@@ -53,6 +55,7 @@ namespace wrench {
              * @return false if the daemon should terminate
              */
             bool Node::processNextMessage() {
+                //TODO: add cpu overhead to... everything
                 S4U_Simulation::computeZeroFlop();
 
                 // Wait for a message
@@ -71,39 +74,163 @@ namespace wrench {
                     try {
                         S4U_Mailbox::putMessage(msg->ack_mailbox,
                                                 new ServiceDaemonStoppedMessage(this->getMessagePayloadValue(
-                                                        SimpleStorageServiceMessagePayload::DAEMON_STOPPED_MESSAGE_PAYLOAD)));
+                                                    SimpleStorageServiceMessagePayload::DAEMON_STOPPED_MESSAGE_PAYLOAD)));
                     } catch (std::shared_ptr<NetworkError> &cause) {
                         return false;
                     }
                     return false;
 
-                }  else if (auto msg = dynamic_cast<FileReadRequestMessage *>(message.get())) {
-                    // If it's in cache then we know, otherwise we must "search"
-
-                    if(cached(msg->file)){
+                }  else if (auto msg = dynamic_cast<FileSearchRequestMessage *>(message.get())) {
+                    if(cached(msg->file)){//File Cached
                         auto cached=getCached(msg->file);
-                        shared_ptr<FileLocation> best=cached[0];//use load based search, not just 0
-                        StorageService::readFile(msg->file,best,msg->answer_mailbox,msg->mailbox_to_receive_the_file_content,msg->num_bytes_to_read);
-                    }else{
+                        shared_ptr<FileLocation> best=*cached.begin();//use load based search, not just start
+                        try {
+                            S4U_Mailbox::putMessage(msg->answer_mailbox,
+                                                    new FileSearchAnswerMessage(
+                                                            msg->file,
+                                                            best,
+                                                            true,
+                                                            nullptr,
+                                                            getMessagePayloadValue(MessagePayload::FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD))
+                            );
+                        } catch (std::shared_ptr<NetworkError> &cause) {
+                            throw ExecutionException(cause);
+                        }
+                    }else{//File Not Cached
+                        if(internalStorage&&StorageService::lookupFile(msg->file,FileLocation::LOCATION(internalStorage))){
+                            //File in internal storage
+                            cache.add(msg->file,FileLocation::LOCATION(internalStorage));
+                            try {
+                                S4U_Mailbox::putMessage(msg->answer_mailbox,
+                                                        new FileSearchAnswerMessage(
+                                                                msg->file,
+                                                                FileLocation::LOCATION(internalStorage),
+                                                                true,
+                                                                nullptr,
+                                                                getMessagePayloadValue(MessagePayload::FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD))
+                                );
+                            } catch (std::shared_ptr<NetworkError> &cause) {
+                                throw ExecutionException(cause);
+                            }
+                        }else{//File not in internal storage or cache
+                            if(children.size()>0){//shotgun continued search message to all chldren
+                                shared_ptr<bool> answered=make_shared<bool>(false);
 
+                                for(auto child:children){
+                                    S4U_Mailbox::putMessage(child->mailbox,
[…]
