### 2022-05-12 3a45494
added Naive implimentation of XRootD
