### 2022-05-12 9087463
(#261) Added bottom-level computation in addition to top-level computation. Added unit tests.
