+++ tools/wfcommons/src/WfCommonsWorkflowParser.cpp
@@ -37,7 +37,7 @@ namespace wrench {
         std::set<std::string> ignored_transfer_jobs;
 
         auto workflow = Workflow::createWorkflow();
-        workflow->disableTopLevelDynamicUpdates();
+        workflow->enableTopBottomLevelDynamicUpdates(false);
 
         double flop_rate;
 
@@ -230,8 +230,8 @@ namespace wrench {
             }
         }
         file.close();
-        workflow->enableTopLevelDynamicUpdates();
-        workflow->updateAllTopLevels();
+        workflow->enableTopBottomLevelDynamicUpdates(true);
+        workflow->updateAllTopBottomLevels();
 
         return workflow;
     }
