### 2022-05-12 2570b11
(#261) Added bottom-level computation in addition to top-level computation. Added unit tests.
