### 2022-05-16 7f7d5c6
Added job failure processing to HTCondor Central Manager
