### 2022-05-16 f4f07a0
Added job failure processing to HTCondor Central Manager
