### 2022-05-20 1005f07
updated StorageService API
added object API to storage service
