### 2022-05-20 5fa5859
Fixed test error
Turns out this nullcheck WAS important after all
