+++ src/wrench/services/storage/xrootd/XRootD.cpp
@@ -13,11 +13,11 @@
 namespace wrench {
     namespace XRootD{
 
-        std::shared_ptr<Node> XRootD::createStorageServer(const std::string& hostname,std::set <std::string> path,WRENCH_PROPERTY_COLLECTION_TYPE property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list){
+        std::shared_ptr<Node> XRootD::createStorageServer(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list){
             std::shared_ptr<Node> ret=createNode(hostname);
             ret->makeSupervisor();
             supervisors.push_back(ret);
-            ret->makeFileServer(path,property_list,messagepayload_list);
+            ret->makeFileServer({"/"},property_list,messagepayload_list);
             dataservers.push_back(ret);
             return ret;
         }
@@ -27,9 +27,9 @@ namespace wrench {
             supervisors.push_back(ret);
             return ret;
         }
-        std::shared_ptr<Node> XRootD::createStorageSupervisor(const std::string& hostname,std::set <std::string> path,WRENCH_PROPERTY_COLLECTION_TYPE property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list){
+        std::shared_ptr<Node> XRootD::createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list){
             auto ret=createNode(hostname);
-            ret->makeFileServer(path,property_list,messagepayload_list);
+            ret->makeFileServer({"/"},property_list,messagepayload_list);
             dataservers.push_back(ret);
             return ret;
         }
