### 2022-05-28 7fbcc01
added object api to simulation
I missed this on the first pass
