### 2022-05-28 5af1423
added object api to simulation
I missed this on the first pass
