+++ include/wrench/services/storage/xrootd/Node.h
@@ -45,6 +45,11 @@ namespace wrench {
             void readFile(std::shared_ptr<DataFile>file);
             void readFile(std::shared_ptr<DataFile>file, double num_bytes);
 
+            std::string getMountPoint() override;
+            std::set<std::string> virgetMountPoints() override;
+            bool hasMultipleMountPoints() override;
+            bool hasMountPoint(const std::string &mp) override;
+
             bool cached(shared_ptr<DataFile> file);
             std::set<std::shared_ptr<FileLocation>> getCached(shared_ptr<DataFile> file);
             int addChild(std::shared_ptr<Node> child);
