### 2022-06-01 253da65
added add function that takes an already wrapped smart pointer
