+++ include/wrench/services/storage/StorageService.h
@@ -45,7 +45,6 @@ namespace wrench {
         std::map<std::string, double> getFreeSpace();
 
         std::map<std::string, double> getTotalSpace();
-
         virtual std::string getMountPoint();
         virtual std::set<std::string> getMountPoints();
         virtual bool hasMultipleMountPoints();
