+++ src/wrench/services/storage/storage_helper_classes/LogicalFileSystem.cpp
@@ -10,6 +10,7 @@
 #include <wrench-dev.h>
 #include <wrench/services/storage/storage_helpers/LogicalFileSystem.h>
 
+#include <limits>
 #include <utility>
 
 WRENCH_LOG_CATEGORY(wrench_core_logical_file_system, "Log category for Logical File System");
@@ -23,27 +24,31 @@ namespace wrench {
      * @brief Constructor
      * @param hostname: the host on which the file system is located
      * @param storage_service: the storage service this file system is for
-     * @param mount_point: the mount point
+     * @param mount_point: the mount point, defaults to /dev/null
      */
     LogicalFileSystem::LogicalFileSystem(const std::string &hostname, StorageService *storage_service, std::string mount_point) {
-        mount_point = FileLocation::sanitizePath("/" + mount_point + "/");
-
         if (storage_service == nullptr) {
             throw std::invalid_argument("LogicalFileSystem::LogicalFileSystem(): nullptr storage_service argument");
         }
-        // Check validity
-        if (not S4U_Simulation::hostHasMountPoint(hostname, mount_point)) {
-            throw std::invalid_argument("LogicalFileSystem::LogicalFileSystem(): Host " +
-                                        hostname + " does not have a disk mounted at " + mount_point);
-        }
-
         this->hostname = hostname;
         this->storage_service = storage_service;
-        this->mount_point = mount_point;
         this->content["/"] = {};
-        this->total_capacity = S4U_Simulation::getDiskCapacity(hostname, mount_point);
+
         this->occupied_space = 0;
         this->initialized = false;
+        if mount_point==DEV_NULL)
+            devnull=true;
+            this->total_capacity = std::numeric_limits<double>::infinity();
+        }else{
+            mount_point = FileLocation::sanitizePath("/" + mount_point + "/");
+            // Check validity
+            if (not S4U_Simulation::hostHasMountPoint(hostname, mount_point)) {
+                throw std::invalid_argument("LogicalFileSystem::LogicalFileSystem(): Host " +
+                                            hostname + " does not have a disk mounted at " + mount_point);
+            }
+            this->total_capacity = S4U_Simulation::getDiskCapacity(hostname, mount_point);
+        }
+        this->mount_point = mount_point;
     }
 
 
@@ -52,7 +57,6 @@ namespace wrench {
      */
     void LogicalFileSystem::init() {
         // Check uniqueness
-
         auto lfs = LogicalFileSystem::mount_points.find(this->hostname + ":" + this->mount_point);
 
         if (lfs != LogicalFileSystem::mount_points.end()) {
@@ -74,6 +78,9 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     void LogicalFileSystem::createDirectory(const std::string &absolute_path) {
+        if(devnull) {
+            return ;
+        }
         assertInitHasBeenCalled();
         assertDirectoryDoesNotExist(absolute_path);
         this->content[absolute_path] = {};
@@ -85,6 +92,9 @@ namespace wrench {
  * @return true if the directory exists
  */
     bool LogicalFileSystem::doesDirectoryExist(const std::string &absolute_path) {
+        if(devnull) {
+            return false;
+        }
         assertInitHasBeenCalled();
         return (this->content.find(absolute_path) != this->content.end());
     }
@@ -97,6 +107,9 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     bool LogicalFileSystem::isDirectoryEmpty(const std::string &absolute_path) {
+        if(devnull) {
+            return true;
+        }
         assertInitHasBeenCalled();
[…]
