### 2022-06-02 5994ffe
fixed a few errors with the new /dev/null
