### 2022-06-02 b2c82b0
fixed a few errors with the new /dev/null
