### 2022-06-08 5dcddc5
set properties
