+++ src/wrench/services/storage/xrootd/XRootDMessagePayload.cpp
@@ -11,11 +11,7 @@ namespace wrench {
     namespace XRootD{
 
 
-        SET_MESSAGEPAYLOAD_NAME(MessagePayload,FILE_LOOKUP_REQUEST_MESSAGE_PAYLOAD);
 
-        SET_MESSAGEPAYLOAD_NAME(MessagePayload,FILE_DELETE_REQUEST_MESSAGE_PAYLOAD);
-
-        SET_MESSAGEPAYLOAD_NAME(MessagePayload,FILE_READ_REQUEST_MESSAGE_PAYLOAD);
         SET_MESSAGEPAYLOAD_NAME(MessagePayload,FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD);
 
         SET_MESSAGEPAYLOAD_NAME(MessagePayload,CONTINUE_SEARCH);
