+++ include/wrench/services/storage/xrootd/Node.h
@@ -36,7 +36,9 @@ namespace wrench {
                     {MessagePayload::UPDATE_CACHE, 1024},
                     {MessagePayload::CONTINUE_SEARCH, 1024},
                     {MessagePayload::FILE_READ_REQUEST_MESSAGE_PAYLOAD, 1024},
-                    {MessagePayload::FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD, 1024}
+                    {MessagePayload::FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD, 1024},
+                    {MessagePayload::CACHE_ENTRY,1024},
+                    {MessagePayload::FILE_DELETE_REQUEST_MESSAGE_PAYLOAD,1024}
             };
         public:
             //XRootD* getMetavisor();
