+++ test/services/helper_services/action_executor/KillFailureActionExecutorTest.cpp
@@ -334,7 +334,7 @@ void KillFailActionExecutorTest::do_ActionExecutorKillFailTest_test(double sleep
     char **argv = (char **) calloc(argc, sizeof(char *));
     argv[0] = strdup("unit_test");
     argv[1] = strdup("--wrench-host-shutdown-simulation");
-    //    argv[2] = strdup("--wrench-full-log");
+//    argv[2] = strdup("--wrench-full-log");
 
     simulation->init(&argc, argv);
 
