### 2022-06-14 cee211b
Committing clang-format changes
