### 2022-06-14 685bd32
Committing clang-format changes
