+++ src/wrench/services/storage/simple/SimpleStorageService.cpp
@@ -221,7 +221,7 @@ namespace wrench {
 
         } else if (auto msg = dynamic_cast<StorageServiceFileReadRequestMessage *>(message.get())) {
             return processFileReadRequest(msg->file, msg->location, msg->num_bytes_to_read, msg->answer_mailbox,
-                                          msg->mailbox_to_receive_the_file_content, msg->buffer_size);
+                                          msg->mailbox_to_receive_the_file_content);
 
         } else if (auto msg = dynamic_cast<StorageServiceFileCopyRequestMessage *>(message.get())) {
             return processFileCopyRequest(msg->file, msg->src, msg->dst, msg->answer_mailbox);
@@ -361,8 +361,7 @@ namespace wrench {
                                                       const std::shared_ptr<FileLocation> &location,
                                                       double num_bytes_to_read,
                                                       simgrid::s4u::Mailbox *answer_mailbox,
-                                                      simgrid::s4u::Mailbox *mailbox_to_receive_the_file_content,
-                                                      unsigned long buffer_size) {
+                                                      simgrid::s4u::Mailbox *mailbox_to_receive_the_file_content) {
         // Figure out whether this succeeds or not
         std::shared_ptr<FailureCause> failure_cause = nullptr;
 
@@ -395,6 +394,7 @@ namespace wrench {
                         location,
                         success,
                         failure_cause,
+                        buffer_size,
                         this->getMessagePayloadValue(
                                 SimpleStorageServiceMessagePayload::FILE_READ_ANSWER_MESSAGE_PAYLOAD)));
 
