+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -53,6 +53,7 @@ namespace wrench {
          * @return the child id of the new node, or -1 if the supervisor already has 64
          */
             int Node::addChild(std::shared_ptr<Node> child){
+                child->supervisor=this;
                 if(children.size()<64){
                     children.push_back(child);
                     return children.size()-1;
@@ -113,12 +114,12 @@ namespace wrench {
                 //S4U_Simulation::compute(flops);
                 S4U_Simulation::computeZeroFlop();
                 // Wait for a message
-                std::shared_ptr<SimulationMessage> message = nullptr;
+                std::shared_ptr <SimulationMessage> message = nullptr;
 
                 S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::MESSAGE_OVERHEAD));
                 try {
                     message = S4U_Mailbox::getMessage(this->mailbox);
-                } catch (std::shared_ptr<NetworkError> &cause) {
+                } catch (std::shared_ptr <NetworkError> &cause) {
                     WRENCH_INFO("Got a network error while getting some message... ignoring");
                     return true;// oh well
                 }
@@ -128,75 +129,92 @@ namespace wrench {
                     try {
                         S4U_Mailbox::putMessage(msg->ack_mailbox,
                                                 new ServiceDaemonStoppedMessage(this->getMessagePayloadValue(
-                                                    SimpleStorageServiceMessagePayload::DAEMON_STOPPED_MESSAGE_PAYLOAD)));
-                    } catch (std::shared_ptr<NetworkError> &cause) {
+                                                        SimpleStorageServiceMessagePayload::DAEMON_STOPPED_MESSAGE_PAYLOAD)));
+                    } catch (std::shared_ptr <NetworkError> &cause) {
                         return false;
                     }
                     return false;
 
-                }  else if (auto msg = dynamic_cast<FileSearchRequestMessage *>(message.get())) {
+                } else if (auto msg = dynamic_cast<StorageServiceFileReadRequestMessage *>(message.get())) {
 
                     S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::CACHE_LOOKUP_OVERHEAD));
-                    if(cached(msg->file)){//File Cached
-                        auto cached=getCached(msg->file);
-                        shared_ptr<FileLocation> best=*cached.begin();//use load based search, not just start
+                    if (cached(msg->file)) {//File Cached
+                        auto cacheCopies = getCached(msg->file);
+                        shared_ptr <FileLocation> best = *cacheCopies.begin();//TODO use load based search, not just start
+
                         try {
-                            S4U_Mailbox::putMessage(msg->answer_mailbox,
-                                                    new FileSearchAnswerMessage(
+                            S4U_Mailbox::putMessage(best->getStorageService()->mailbox,
+                                                    new StorageServiceFileReadRequestMessage(
+                                                            msg->answer_mailbox,
+                                                            msg->mailbox_to_receive_the_file_content,
                                                             msg->file,
                                                             best,
-                                                            true,
-                                                            nullptr,
-                                                            getMessagePayloadValue(MessagePayload::FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD))
+                                                            msg->num_bytes_to_read,
+                                                            getMessagePayloadValue(
+                                                                    MessagePayload::FILE_SEARCH_ANSWER_MESSAGE_PAYLOAD))
                             );
-                        } catch (std::shared_ptr<NetworkError> &cause) {
+                        } catch (std::shared_ptr <NetworkError> &cause) {
                             throw ExecutionException(cause);
                         }
-                    }else{//File Not Cached
-                        if(internalStorage&&StorageService::lookupFile(msg->file,FileLocation::LOCATION(internalStorage))){
[…]
