### 2022-06-16 3d48c3a
Caching invalid message_payload index error (for <0 index)
