### 2022-06-16 40a2237
Caught an out-of-range exception when looking up message payload name
