+++ src/wrench/services/Service.cpp
@@ -154,14 +154,16 @@ namespace wrench {
      * @throw std::invalid_argument
      */
     double Service::getMessagePayloadValue(WRENCH_MESSAGEPAYLOAD_TYPE message_payload) {
-        if (message_payload < 0) {
-            throw std::invalid_argument("Service::getMessagePayloadValue(): invalid message payload specification");
-        }
         if (this->messagepayload_list.find(message_payload) == this->messagepayload_list.end()) {
-            throw std::invalid_argument(
+            try {
+                throw std::invalid_argument(
                     "Service::getMessagePayloadValue(): Cannot find value for message_payload " +
                     ServiceMessagePayload::translatePayloadType(message_payload) +
                     " (perhaps a derived service class does not provide a default value?)");
+            } catch (std::out_of_range &e) {
+                throw std::invalid_argument(
+                        "Service::getMessagePayloadValue(): invalid message_payload index");
+            }
         }
         return this->messagepayload_list[message_payload];
     }
