+++ include/wrench/services/storage/simple/SimpleStorageService.h
@@ -80,7 +80,8 @@ namespace wrench {
         ~SimpleStorageService() override;
 
         void cleanup(bool has_returned_from_main, int return_value) override;
-        
+        double getLoad() override;
+        double countRunningFileTransferThreads();
         /***********************/
         /** \endcond          **/
         /***********************/
