+++ src/wrench/services/storage/simple/SimpleStorageService.cpp
@@ -668,5 +668,20 @@ namespace wrench {
         this->getPropertyValueAsUnsignedLong(SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS);
         this->getPropertyValueAsUnsignedLong(SimpleStorageServiceProperty::BUFFER_SIZE);
     }
+    /**
+         * @brief Get number of File Transfer Threads that are currently running or are pending
+         * @return The number of threads
+         */
+    double SimpleStorageService::countRunningFileTransferThreads() {
+        return running_file_transfer_threads.size()+pending_file_transfer_threads.size();
 
+    }
+        /**
+         * @brief Get the load (number of concurrent reads) on the storage service
+         * @return the load on the service
+         */
+    double SimpleStorageService::getLoad() {
+        return countRunningFileTransferThreads();
+
+    }
 };// namespace wrench
