+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -23,43 +23,45 @@
 WRENCH_LOG_CATEGORY(wrench_core_xrootd_data_server,
                     "Log category for XRootD");
 namespace wrench {
-    namespace XRootD{
-            int Node::main() {
-                //TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_CYAN);
+    namespace XRootD {
+        int Node::main() {
+            //TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_CYAN);
 
-                // Start file storage server
-                //if(internalStorage){
-                //    internalStorage->start(internalStorage,true,true);
-                //}
+            // Start file storage server
+            //if(internalStorage){
+            //    internalStorage->start(internalStorage,true,true);
+            //}
 
-                std::string message = "XRootD Node " + this->getName() + "  starting on host " + this->getHostname();
-                WRENCH_INFO("%s", message.c_str());
+            std::string message =
+                    "XRootD Node " + this->getName() + "  starting on host " + this->getHostname();WRENCH_INFO("%s",
+                                                                                                               message.c_str());
 
 
-                /** Main loop **/
-                while (this->processNextMessage()) {
+            /** Main loop **/
+            while (this->processNextMessage()) {
 
-                }
-
-                WRENCH_INFO("XRootD Node %s on host %s cleanly terminating!",
-                            this->getName().c_str(),
-                            S4U_Simulation::getHostName().c_str());
-
-                return 0;
             }
+
+            WRENCH_INFO("XRootD Node %s on host %s cleanly terminating!",
+                        this->getName().c_str(),
+                        S4U_Simulation::getHostName().c_str());
+
+            return 0;
+        }
+
         /**
          * @brief Adds a child node to an XRootD supervisor
          * @param child: the new child to add
          * @return the child id of the new node, or -1 if the supervisor already has 64
          */
-            int Node::addChild(std::shared_ptr<Node> child){
-                child->supervisor=this;
-                if(children.size()<64){
-                    children.push_back(child);
-                    return children.size()-1;
-                }
-                return -1;
+        int Node::addChild(std::shared_ptr<Node> child) {
+            child->supervisor = this;
+            if (children.size() < 64) {
+                children.push_back(child);
+                return children.size() - 1;
             }
+            return -1;
+        }
 //            /**
 //             * @brief Get mount point of underlying storage service
 //             * @return the (sole) mount point of the service if there is one, and null otherwise
@@ -103,128 +105,150 @@ namespace wrench {
 //                return false;
 //
 //            }
-            /**
-             * @brief Process a received control message
-             *
-             * @return false if the daemon should terminate
-             */
+        /**
+         * @brief Process a received control message
+         *
+         * @return false if the daemon should terminate
+         */
 
-            bool Node::processNextMessage() {
-                //TODO: add cpu overhead to... everything
-                //S4U_Simulation::compute(flops);
-                S4U_Simulation::computeZeroFlop();
-                // Wait for a message
-                std::unique_ptr <SimulationMessage> message = nullptr;
+        bool Node::processNextMessage() {
+            //TODO: add cpu overhead to... everything
+            //S4U_Simulation::compute(flops);
+            S4U_Simulation::computeZeroFlop();
+            // Wait for a message
+            std::unique_ptr<SimulationMessage> message = nullptr;
 
-                S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::MESSAGE_OVERHEAD));
+            S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::MESSAGE_OVERHEAD));
+            try {
+                message = S4U_Mailbox::getMessage(this->mailbox);
[…]
