+++ conf.py
@@ -1,0 +1,8 @@
+breathe_projects = {
+    "user": "../../docs/xrootd/user/xml/",
+    "developer": "../../docs/xrootd/developer/xml/",
+    "internal": "../../docs/xrootd/internal/xml/",
+}
+
+version = 'xrootd'
+release = 'xrootd'
