+++ src/wrench/services/storage/xrootd/XRootD.cpp
@@ -12,7 +12,14 @@
 #include <wrench/services/storage/xrootd/XRootD.h>
 namespace wrench {
     namespace XRootD{
-
+        /**
+        * @brief Create a new XRootD Node that is both a storage server and a supervisor
+        * @param hostname: the name of the host on which the service and its storage service should run
+        * @param property_list: The property list to use for the Node AND the underlying storage server
+        * @param messagepayload_list: The message payload list to use for the Node AND the underlying storage server
+        *
+        * @return a shared pointer to the newly created Node
+        */
         std::shared_ptr<Node> XRootD::createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list){
             std::shared_ptr<Node> ret=createNode(hostname);
             ret->makeSupervisor();
@@ -22,12 +29,28 @@ namespace wrench {
             dataservers.push_back(ret);
             return ret;
         }
+        /**
+        * @brief Create a new XRootD Node that is a supervisor
+        * @param hostname: the name of the host on which the service should run
+        * @param property_list: The property list to use for the supervisor
+        * @param messagepayload_list: The message payload list to use for supervisor
+        *
+        * @return a shared pointer to the newly created Node
+        */
         std::shared_ptr<Node> XRootD::createSupervisor(const std::string& hostname){
             std::shared_ptr<Node> ret=createNode(hostname);
             ret->makeSupervisor();
             supervisors.push_back(ret);
             return ret;
         }
+        /**
+        * @brief Create a new XRootD Node that is a leaf storage server
+        * @param hostname: the name of the host on which the service and its storage service should run
+        * @param property_list: The property list to use for the host AND the underlying storage server
+        * @param messagepayload_list: The message payload list to use for the host AND the underlying storage server
+        *
+        * @return a shared pointer to the newly created Node
+        */
         std::shared_ptr<Node> XRootD::createStorageServer(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list){
             auto ret=createNode(hostname);
             ret->makeFileServer({"/"},property_list,messagepayload_list);
@@ -35,7 +58,12 @@ namespace wrench {
             dataservers.push_back(ret);
             return ret;
         }
-
+        /**
+        * @brief Create a new blank XRootD Node.  The newly created Node is neither a supervisor nor a storage server.  This function should not be called directly.
+        * @param hostname: the name of the host on which the service and its storage service should run
+        *
+        * @return a shared pointer to the newly created Node
+        */
         std::shared_ptr<Node> XRootD::createNode(const std::string& hostname){
             std::shared_ptr<Node> ret= make_shared<Node>(hostname);
             ret->metavisor=this;
@@ -43,14 +71,44 @@ namespace wrench {
             simulation->add(ret);
             return ret;
         }
+        /**
+        * @brief Meta operation to get all File servers in the federation that have the file specified, intended for advanced file search simulation optimization.
+        * @param file: A shared pointer to the file to search for
+        *
+        * @return vector of Node pointers that are all File Servers with the file
+        */
         std::vector<std::shared_ptr<Node>> XRootD::getFileNodes(std::shared_ptr<DataFile> file){
             return files[file];
 
         }
+        /**
+        * @brief get the size of the XRootD federation
+        *
+        * @return the number of nodes in the federation
+        */
         unsigned int XRootD::size(){
             return nodes.size();
 
         }
+        /**
[…]
