+++ conf.py
@@ -1,8 +1,0 @@
-breathe_projects = {
-    "user": "../../docs/xrootd/user/xml/",
-    "developer": "../../docs/xrootd/developer/xml/",
-    "internal": "../../docs/xrootd/internal/xml/",
-}
-
-version = 'xrootd'
-release = 'xrootd'
