+++ include/wrench/services/storage/xrootd/Cache.h
@@ -16,17 +16,17 @@
 #include <set>
 #include "wrench/data_file/DataFile.h"
 #include "wrench/services/storage/storage_helpers/FileLocation.h"
-
+/***********************/
+/** \cond INTERNAL     */
+/***********************/
 namespace wrench{
     namespace XRootD{
         class Cache{
         private:
-            /***********************/
-            /** \cond INTERNAL     */
-            /***********************/
 
-            /** @brief The internal cache datastructure, currently just a map of data files pointers to a set of file locations.  In future, this will contain timestamps */
-            std::unordered_map< std::shared_ptr<DataFile>, std::set<std::shared_ptr<FileLocation>>> cache;//probiably change the payload of this to an object containing the file location AND its queue time stamp
+
+            /** @brief The internal cache data structure, currently just a map of data files pointers to a set of file locations.  In future, this will contain timestamps */
+            std::unordered_map< std::shared_ptr<DataFile>, std::set<std::shared_ptr<FileLocation>>> cache;//probably change the payload of this to an object containing the file location AND its queue time stamp
         public:
             bool isCached(std::shared_ptr<DataFile> file);
             void add(std::shared_ptr<DataFile> file,std::shared_ptr<FileLocation> location);
@@ -36,11 +36,12 @@ namespace wrench{
 
             std::set<std::shared_ptr<FileLocation>> operator[](std::shared_ptr<DataFile> file);
             void remove(std::shared_ptr<DataFile> file);
-            /***********************/
-            /** \endcond           */
-            /***********************/
+
         };
 
     }
 }
+/***********************/
+/** \endcond           */
+/***********************/
 #endif//WRENCH_CACHE_H
