+++ src/wrench/services/storage/xrootd/XRootD.cpp
@@ -15,13 +15,15 @@ namespace wrench {
         /**
         * @brief Create a new XRootD Node that is both a storage server and a supervisor
         * @param hostname: the name of the host on which the service and its storage service should run
-        * @param storage_property_list: The property list to use for the Node AND the underlying storage server
-        * @param storage_messagepayload_list: The message payload list to use for the Node AND the underlying storage server
+        * @param storage_property_list: The property list to use for the underlying storage server
+        * @param storage_messagepayload_list: The message payload list to use for the underlying storage server
+        * @param node_property_list: The property list to use for the new Node, defaluts to {}
+        * @param node_messagepayload_list: The message payload list to use for the new Node, defaults to {}
         *
         * @return a shared pointer to the newly created Node
         */
-        std::shared_ptr<Node> XRootD::createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list){
-            std::shared_ptr<Node> ret=createNode(hostname);
+        std::shared_ptr<Node> XRootD::createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list){
+            std::shared_ptr<Node> ret=createNode(hostname,node_property_list,node_messagepayload_list);
             ret->makeSupervisor();
             supervisors.push_back(ret);
             ret->makeFileServer({"/"},storage_property_list,storage_messagepayload_list);
@@ -34,11 +36,13 @@ namespace wrench {
         * @param hostname: the name of the host on which the service should run
         * @param property_list: The property list to use for the supervisor
         * @param messagepayload_list: The message payload list to use for supervisor
+        * @param node_property_list: The property list to use for the new Node, defaluts to {}
+        * @param node_messagepayload_list: The message payload list to use for the new Node, defaults to {}
         *
         * @return a shared pointer to the newly created Node
         */
-        std::shared_ptr<Node> XRootD::createSupervisor(const std::string& hostname){
-            std::shared_ptr<Node> ret=createNode(hostname);
+        std::shared_ptr<Node> XRootD::createSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list){
+            std::shared_ptr<Node> ret=createNode(hostname,node_property_list,node_messagepayload_list);
             ret->makeSupervisor();
             supervisors.push_back(ret);
             return ret;
@@ -48,11 +52,13 @@ namespace wrench {
         * @param hostname: the name of the host on which the service and its storage service should run
         * @param storage_property_list: The property list to use for the underlying storage server
         * @param storage_messagepayload_list: The message payload list to use for the underlying storage server
+        * @param node_property_list: The property list to use for the new Node, defaluts to {}
+        * @param node_messagepayload_list: The message payload list to use for the new Node, defaults to {}
         *
         * @return a shared pointer to the newly created Node
         */
[…]
