+++ src/wrench/services/storage/xrootd/Cache.cpp
@@ -16,7 +16,19 @@ namespace wrench{
          * @return true if the file is cached and if timestamp (not implemented) is valid, false otherwise
          */
         bool Cache::isCached(std::shared_ptr<DataFile> file) {
-            return cache.find(file)!=cache.end();//once timestamps are implemented also check the file is still in cache  it (remove if not)
+            double earliestAllowedTime = wrench::S4U_Simulation::getClock()-maxCacheTime;
+            auto entries=cache[file];
+            //after getting all possible cache entries, loop through them "all" and check the timestamps, removing any that fail.  If you find even 1 success, return true and stop cleaning.
+            for(auto ittr=entries.begin();ittr!=entries.end();){//intentionaly blank 3rd term
+                auto entry=*ittr;
+                if(entry.second<earliestAllowedTime){
+                    ittr=entries.erase(ittr);
+                }else{
+                    return true;
+                }
+            }
+            return false;
+
         }
         /**
          * @brief Add a file to the cache
@@ -24,18 +36,19 @@ namespace wrench{
          * @param location: The location to add to the cache
          */
         void Cache::add(std::shared_ptr<DataFile> file,std::shared_ptr<FileLocation> location){
-            wrench::S4U_Simulation::getClock();
-            cache[file].insert(location);
+            double currentSimTime = wrench::S4U_Simulation::getClock();
+            cache[file][location]=currentSimTime;
         }
+
         /**
         * @brief Add a file to the cache
         * @param file: The file to add to the cache
         * @param locations: The locations to add to the cache
         */
         void Cache::add(std::shared_ptr<DataFile> file, std::set<std::shared_ptr<FileLocation>> locations){
-            wrench::S4U_Simulation::getClock();
+            double currentSimTime = wrench::S4U_Simulation::getClock();
             for(auto location:locations){
-                cache[file].insert(location);
+                cache[file][location]=currentSimTime;
             }
         }
         /**
@@ -44,7 +57,20 @@ namespace wrench{
          * @return the set of valid cached copies.  (empty set if not found)
          */
         std::set<std::shared_ptr<FileLocation>> Cache::get(std::shared_ptr<DataFile> file){
-            return cache[file];
+            double earliestAllowedTime = wrench::S4U_Simulation::getClock()-maxCacheTime;
+            auto entries=cache[file];
+            std::set<std::shared_ptr<FileLocation>> ret;
+            for(auto ittr=entries.begin();ittr!=entries.end();){//intentionaly blank 3rd term
+                auto entry=*ittr;
+                if(entry.second<earliestAllowedTime){
+                    ittr=entries.erase(ittr);
+                }else{
+                    ret.insert(entry.first);
+                    ittr++;
+
+                }
+            }
+            return ret;
 
         }/**
          * @brief get all valid cached copies.
@@ -61,6 +87,34 @@ namespace wrench{
 
         void Cache::remove(std::shared_ptr<DataFile> file){
             cache.erase(file);
+        }
+        /**
+         * @brief Deap Clean the cache of outdated entries.  Possibly slow, but if the cache size is getting out of hand, this can help.
+         * @param file: The file to check the cache for
+         */
+
+        void Cache::clean(){
+            double earliestAllowedTime = wrench::S4U_Simulation::getClock()-maxCacheTime;
+            //itterate all files
+            for(auto ettr=cache.begin();ettr!=cache.end();) {//intentionaly blank 3rd term
+                auto entries = ettr->second;
+                //iterate all locations
+                for (auto ittr = entries.begin(); ittr != entries.end();) {//intentionaly blank 3rd term
+                    auto entry = *ittr;
+                    //check all timestamps
+                    if (entry.second < earliestAllowedTime) {
+                        ittr = entries.erase(ittr);
[…]
