+++ src/wrench/services/storage/storage_helper_classes/LogicalFileSystem.cpp
@@ -342,4 +342,5 @@ namespace wrench {
             return;
         }
     }
+
 }// namespace wrench
