+++ include/wrench/services/storage/xrootd/XRootD.h
@@ -42,10 +42,10 @@ namespace wrench {
              * @param property_values: The property values that should be used to overwrite the defaults of all Nodes (defaults to none) (unless otherwise specified)
              * @param messagepayload_values: The message paylaod values that should be used to overwrite the defaults of all Nodes (defaults to none) (unless otherwise specified)
              */
-            XRootD(std::shared_ptr<Simulation>  simulation,WRENCH_PROPERTY_COLLECTION_TYPE property_values={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_values={}):simulation(simulation),property_values(property_values),messagepayload_values(messagepayload_values){}
-            std::shared_ptr<Node> createStorageServer(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list);
-            std::shared_ptr<Node> createSupervisor(const std::string& hostname);
-            std::shared_ptr<Node> createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list, WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list);
+            XRootD(std::shared_ptr<Simulation>  simulation,WRENCH_PROPERTY_COLLECTION_TYPE property_values={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_values={}):property_values(property_values),messagepayload_values(messagepayload_values),simulation(simulation){}
+            std::shared_ptr<Node> createStorageServer(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
+            std::shared_ptr<Node> createSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
+            std::shared_ptr<Node> createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list, WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
             /***********************/
             /** \cond DEVELOPER    */
             /***********************/
@@ -65,7 +65,7 @@ namespace wrench {
             /***********************/
             friend Node;
             std::vector<std::shared_ptr<Node>> getFileNodes(std::shared_ptr<DataFile> file);
-            std::shared_ptr<Node> createNode(const std::string& hostname);
+            std::shared_ptr<Node> createNode(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE property_list_override,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_list_override);
             /** @brief All nodes that are connected to this XRootD data Federation */
             std::vector<std::shared_ptr<Node>> nodes;
             /** @brief All nodes in the XRootD Federation that have and internal file server */
