### 2022-06-28 99dd673
added way to remove files from XRootD registry
