### 2022-06-28 8ca07ef
added way to remove files from XRootD registry
