+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -38,6 +38,7 @@ namespace wrench {
             std::string message =
                     "XRootD Node " + this->getName() + "  starting on host " + this->getHostname();WRENCH_INFO("%s",
                                                                                                                message.c_str());
+            reduced= getPropertyValueAsBoolean(Property::REDUCED_SIMULATION);
             /** Main loop **/
             while (this->processNextMessage()) {
 
@@ -128,7 +129,139 @@ namespace wrench {
             }
 
             WRENCH_DEBUG("Got a [%s] message", message->getName().c_str());
-            if (auto msg = dynamic_cast<ServiceStopDaemonMessage *>(message.get())) {
+            if(reduced){//handle the 2 Advancecd Messages if reduced simulation is enabled for this node
+                if (auto msg = dynamic_cast<AdvancedContinueSearchMessage *>(message.get())) {
+                    S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::CACHE_LOOKUP_OVERHEAD));
+                    if(cached(msg->file)){//caching is even more efficient than advanced search
+                        auto cached = getCached(msg->file);
+                        try {
+                            S4U_Mailbox::putMessage(supervisor->mailbox,
+                                                    new UpdateCacheMessage(
+                                                            msg->answer_mailbox,
+                                                            msg->original,
+                                                            msg->node,
+                                                            msg->file,
+                                                            cached,
+                                                            getMessagePayloadValue(MessagePayload::UPDATE_CACHE) +
+                                                            getMessagePayloadValue(MessagePayload::CACHE_ENTRY) *
+                                                            cached.size(),
+                                                            msg->answered
+                                                    )
+                            );
+                        } catch (std::shared_ptr<NetworkError> &cause) {
+                            throw ExecutionException(cause);
+                        }
+                    }else {
+                        if(children.size()>0) {
+
+                            map<Node *, vector<stack<Node *>>> splitStacks = splitStack(msg->searchStack);
+                            S4U_Simulation::compute(
+                                    this->getPropertyValueAsDouble(Property::SEARCH_BROADCAST_OVERHEAD));
+                                for (auto entry: splitStacks) {
+                                    if (entry.first == this) {//this node was the target
+                                        if (internalStorage &&//check the storage, it SHOULD be there, but we should check still
+                                            StorageService::lookupFile(msg->file, FileLocation::LOCATION(internalStorage))) {
+                                            //File in internal storage
+                                            cache.add(msg->file, FileLocation::LOCATION(internalStorage));
+                                            try {
+                                                S4U_Mailbox::putMessage(supervisor->mailbox,
+                                                                        new UpdateCacheMessage(
+                                                                                msg->answer_mailbox,
+                                                                                msg->original,
+                                                                                msg->node,
+                                                                                msg->file,
+                                                                                set<std::shared_ptr<FileLocation
+                                                                                >>{FileLocation::LOCATION(internalStorage)},
+                                                                                getMessagePayloadValue(MessagePayload::UPDATE_CACHE) +
+                                                                                getMessagePayloadValue(MessagePayload::CACHE_ENTRY),
+                                                                                msg->answered
+                                                                        )
+                                                );
+                                            } catch (std::shared_ptr<NetworkError> &cause) {
+                                                throw ExecutionException(cause);
+                                            }
+                                        }
+                                    } else {
+                                        try {
+                                            S4U_Mailbox::putMessage(entry.first->mailbox,
[…]
