+++ src/wrench/services/storage/xrootd/XRootDMessage.cpp
@@ -75,7 +75,13 @@ namespace wrench {
         * @param other: The message to copy.
         */
         UpdateCacheMessage::UpdateCacheMessage(UpdateCacheMessage& other):Message(other.payload),answer_mailbox(other.answer_mailbox),original(other.original),file(other.file),locations(other.locations),node(other.node),answered(other.answered){}
-
+        /**
+        * @brief Constructor
+        * @param file: The file to delete.
+        * @param payload: the message size in bytes
+        * @param timeToLive:  The max number of hops this message can take
+        */
+        RippleDelete::RippleDelete(std::shared_ptr<DataFile> file,double payload,int timeToLive):Message(payload),file(file),timeToLive(timeToLive){};
         /**
         * @brief Copy Constructor
         * @param other: The message to copy.
@@ -88,8 +94,56 @@ namespace wrench {
         * @param timeToLive:  The max number of hops this message can take
         */
         RippleDelete::RippleDelete(StorageServiceFileDeleteRequestMessage* other,int timeToLive):Message(other->payload),file(other->file),timeToLive(timeToLive){}
-
-
+        /**
+         * @brief Constructor
+         * @param answer_mailbox: The mailbox the final answer should be sent to
+         * @param original: The original file read request being responded too.  If this is a file locate search, this should be null
+         * @param file: The file to search for
+         * @param node: The node where the search was initiated
+         * @param payload: The message size in bytes
+         * @param answered: A shared boolean for if the answer has been sent to the client.  This should be the same for all messages searching for this request.  Used to prevent the multiple response problem
+         * @param timeToLive: The max number of hops this message can take
+         * @param searchStack:  The available paths to the file
+         */
+        AdvancedContinueSearchMessage::AdvancedContinueSearchMessage(simgrid::s4u::Mailbox *answer_mailbox,std::shared_ptr<StorageServiceFileReadRequestMessage> original,
+                std::shared_ptr<DataFile> file, Node* node, double payload,std::shared_ptr<bool> answered,int timeToLive,std::vector<std::stack<Node*>> searchStack):ContinueSearchMessage(answer_mailbox,original,file,node,payload,answered,timeToLive),searchStack(searchStack){};
+        /**
+        * @brief Pointer Copy Constructor with auxiliary stack
+        * @param other: The message to copy, timeToLive is decremented
+        * @param searchStack:  The available paths to the file
+        */
+        AdvancedContinueSearchMessage::AdvancedContinueSearchMessage(ContinueSearchMessage* toCopy,std::vector<std::stack<Node*>> searchStack): ContinueSearchMessage(toCopy),searchStack(searchStack){};
+        /**
+        * @brief Pointer Copy Constructor
+        * @param other: The message to copy, timeToLive is decremented
+        */
+        AdvancedContinueSearchMessage::AdvancedContinueSearchMessage(AdvancedContinueSearchMessage* toCopy): ContinueSearchMessage(toCopy),searchStack(toCopy->searchStack){};
+        /**
+        * @brief Constructor
+        * @param file: The file to delete.
+        * @param payload: the message size in bytes
+        * @param timeToLive:  The max number of hops this message can take
+        * @param searchStack:  The available paths to the file
+        */
+        AdvancedRippleDelete::AdvancedRippleDelete(std::shared_ptr<DataFile> file,double payload,int timeToLive,std::vector<std::stack<Node*>> searchStack):RippleDelete(file,payload,timeToLive),searchStack(searchStack){}
+        /**
+        * @brief Copy Constructor with auxiliary stack
+        * @param other: The message to copy.
+        * @param searchStack:  The available paths to the file
+        */
[…]
