+++ src/wrench/services/storage/xrootd/XRootDProperty.cpp
@@ -7,5 +7,6 @@ namespace wrench {
         SET_PROPERTY_NAME(Property,SEARCH_BROADCAST_OVERHEAD);
         SET_PROPERTY_NAME(Property,UPDATE_CACHE_OVERHEAD);
         SET_PROPERTY_NAME(Property,CACHE_MAX_LIFETIME);
+        SET_PROPERTY_NAME(Property,REDUCED_SIMULATION);
     }
 }
