### 2022-06-29 0171750
stray {
