+++ include/wrench/services/storage/xrootd/Node.h
@@ -88,6 +88,8 @@ namespace wrench {
         private:
             map<Node*,vector<stack<Node*>>> splitStack(vector<stack<Node*>> searchStack);
             virtual std::shared_ptr<FileLocation> selectBest(std::set<std::shared_ptr<FileLocation>> locations);
+            vector<stack<Node*>> constructFileSearchTree(vector<shared_ptr<Node>>& targets);
+            stack<Node*>constructSearchStack(Node* target);
             //std::shared_ptr<FileLocation> hasFile(shared_ptr<DataFile> file);
 
             bool makeSupervisor();
