### 2022-06-30 8a247c6
theoreticaly advanced search should be good to go
