+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -114,7 +114,6 @@ namespace wrench {
          */
 
         bool Node::processNextMessage() {
-            //TODO: add cpu overhead to... everything
             //S4U_Simulation::compute(flops);
             S4U_Simulation::computeZeroFlop();
             // Wait for a message
@@ -317,9 +316,37 @@ namespace wrench {
                                 this->getPropertyValueAsDouble(Property::SEARCH_BROADCAST_OVERHEAD));
 
                         if (children.size() > 0) {//recursive search
-                            if(reduced){//TODO use optimised search
+                            if(reduced){
+                                shared_ptr<bool> answered = make_shared<bool>(false);
+                                auto targets=metavisor->getFileNodes(msg->file);
+                                auto searchStack=constructFileSearchTree(targets);
+                                map<Node *, vector<stack<Node *>>> splitStacks = splitStack(searchStack);
+                                S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::SEARCH_BROADCAST_OVERHEAD));
+                                for (auto entry: splitStacks) {
+                                    if (entry.first == this) {//this node was the target
+                                        //we shouldnt have to worry about this, it should have been handled earlier.  But just in case, I dont want a rogue search going who knows where
+                                    } else {
+                                        try {
+                                            S4U_Mailbox::putMessage(entry.first->mailbox,
+                                                                    new AdvancedContinueSearchMessage(
+                                                                            msg->answer_mailbox,
+                                                                            make_shared<StorageServiceFileReadRequestMessage>(msg),
+                                                                            msg->file,
+                                                                            this,
+                                                                            getMessagePayloadValue(
+                                                                                    MessagePayload::CONTINUE_SEARCH),
+                                                                            answered,
+                                                                            metavisor->defaultTimeToLive,
+                                                                            entry.second
+                                                                    )
+                                            );
+                                        } catch (std::shared_ptr<NetworkError> &cause) {
+                                            throw ExecutionException(cause);
+                                        }
 
+                                    }
 
+                                }
                             }else {//shotgun continued search message to all children
 
                                 shared_ptr<bool> answered = make_shared<bool>(false);
@@ -474,9 +501,31 @@ namespace wrench {
 
             } else if (auto msg = dynamic_cast<StorageServiceFileDeleteRequestMessage *>(message.get())) {
                 if(reduced) {//TODO use optimised search
-                    metavisor->deleteFile(msg->file);
+
+                    shared_ptr<bool> answered = make_shared<bool>(false);
+                    auto targets=metavisor->getFileNodes(msg->file);
+                    auto searchStack=constructFileSearchTree(targets);
+                    map<Node *, vector<stack<Node *>>> splitStacks = splitStack(searchStack);
+                    S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::SEARCH_BROADCAST_OVERHEAD));
+                    for (auto entry: splitStacks) {
+                        if (entry.first == this) {//this node was the target
+                            //we shouldnt have to worry about this, it should have been handled earlier.  But just in case, I dont want a rogue search going who knows where
+                        } else {
+                            try {
+                                S4U_Mailbox::putMessage(entry.first->mailbox,
+                                                        new AdvancedRippleDelete(
+                                                                msg,
+                                                                metavisor->defaultTimeToLive,
+                                                                entry.second
+                                                        )
[…]
