### 2022-07-12 3a11dd7
Temporarily ifdefed call to simgrid::s4u::thread_execute() due to odd bug on VMs
