+++ test/services/compute_services/bare_metal_compound_jobs/BareMetalComputeServiceOneActionTests.cpp
@@ -253,7 +253,7 @@ void BareMetalComputeServiceOneActionTest::do_BadSetup_test() {
                                                                                              100000.0))},
                                                              "",
                                                              {std::make_pair(
-                                                                     wrench::BareMetalComputeServiceProperty::TASK_STARTUP_OVERHEAD,
+                                                                     wrench::BareMetalComputeServiceProperty::THREAD_STARTUP_OVERHEAD,
                                                                      "-1.0")},
                                                              {})),
                  std::invalid_argument);
