### 2022-07-12 2c85827
Committing clang-format changes
