+++ doc/source/conf.py
@@ -69,3 +69,43 @@ html_theme_options = {
 breathe_default_project = "user"
 
 # The full version, including alpha/beta/rc tags
+breathe_projects = {
+    "user": "../../docs/2.1-dev/user/xml/",
+    "developer": "../../docs/2.1-dev/developer/xml/",
+    "internal": "../../docs/2.1-dev/internal/xml/",
+}
+
+version = '2.1-dev'
+release = '2.1-dev'
+breathe_projects = {
+    "user": "../../docs/2.1-dev/user/xml/",
+    "developer": "../../docs/2.1-dev/developer/xml/",
+    "internal": "../../docs/2.1-dev/internal/xml/",
+}
+
+version = '2.1-dev'
+release = '2.1-dev'
+breathe_projects = {
+    "user": "../../docs/2.1-dev/user/xml/",
+    "developer": "../../docs/2.1-dev/developer/xml/",
+    "internal": "../../docs/2.1-dev/internal/xml/",
+}
+
+version = '2.1-dev'
+release = '2.1-dev'
+breathe_projects = {
+    "user": "../../docs/2.1-dev/user/xml/",
+    "developer": "../../docs/2.1-dev/developer/xml/",
+    "internal": "../../docs/2.1-dev/internal/xml/",
+}
+
+version = '2.1-dev'
+release = '2.1-dev'
+breathe_projects = {
+    "user": "../../docs/2.1-dev/user/xml/",
+    "developer": "../../docs/2.1-dev/developer/xml/",
+    "internal": "../../docs/2.1-dev/internal/xml/",
+}
+
+version = '2.1-dev'
+release = '2.1-dev'
