### 2022-07-13 6c4a8f8
removed catch and throw
