+++ src/wrench/data_file/DataFile.cpp
@@ -39,6 +39,7 @@ namespace wrench {
      * @brief Set the file size (to be used only in very specific
      * cases in which it is guaranteed that changing a file's size after that file
      * has been created is a valid thing to do)
+     * @param s: a size in byte
      */
     void DataFile::setSize(double s) {
         this->size = s;
