### 2022-07-13 283a639
Committing clang-format changes
