### 2022-07-13 7e6908f
SimGrid version fix for thread_execute()
