### 2022-07-14 faef7a3
code cleanup
