### 2022-07-14 3379b03
Committing clang-format changes
