+++ src/wrench/simgrid_S4U_util/S4U_Simulation.cpp
@@ -290,8 +290,8 @@ namespace wrench {
     }
 
 
-    /**
- * @brief Determines whether a host exists for a given hostname
+/**
+ * @brief Determines whether a host exists for a given hostname (does not include VMs)
  * @param hostname: the name of the host
  * @return true or false
  */
@@ -317,7 +317,7 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     unsigned int S4U_Simulation::getHostNumCores(const std::string &hostname) {
-        auto host = simgrid::s4u::Host::by_name_or_null(hostname);
+        auto host = S4U_Simulation::get_host_or_vm_by_name_or_null(hostname);
         if (host == nullptr) {
             throw std::invalid_argument("Unknown hostname " + hostname);
         }
@@ -341,7 +341,7 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     double S4U_Simulation::getHostFlopRate(const std::string &hostname) {
-        auto host = simgrid::s4u::Host::by_name_or_null(hostname);
+        auto host = S4U_Simulation::get_host_or_vm_by_name_or_null(hostname);
         if (host == nullptr) {
             throw std::invalid_argument("Unknown hostname " + hostname);
         }
@@ -357,7 +357,8 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     bool S4U_Simulation::isHostOn(const std::string &hostname) {
-        auto host = simgrid::s4u::Host::by_name_or_null(hostname);
+
+        auto host = S4U_Simulation::get_host_or_vm_by_name_or_null(hostname);
         if (host == nullptr) {
             throw std::invalid_argument("Unknown hostname " + hostname);
         }
@@ -372,7 +373,8 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     void S4U_Simulation::turnOffHost(const std::string &hostname) {
-        auto host = simgrid::s4u::Host::by_name_or_null(hostname);
+
+        auto host = S4U_Simulation::get_host_or_vm_by_name_or_null(hostname);
         if (host == nullptr) {
             throw std::invalid_argument("Unknown hostname " + hostname);
         }
@@ -387,7 +389,8 @@ namespace wrench {
  * @throw std::invalid_argument
  */
     void S4U_Simulation::turnOnHost(const std::string &hostname) {
-        auto host = simgrid::s4u::Host::by_name_or_null(hostname);
+
+        auto host = S4U_Simulation::get_host_or_vm_by_name_or_null(hostname);
         if (host == nullptr) {
             throw std::invalid_argument("Unknown hostname " + hostname);
         }
@@ -518,12 +521,13 @@ namespace wrench {
         mount_point = FileLocation::sanitizePath(mount_point);
 
         WRENCH_DEBUG("Writing %lf bytes to disk %s:%s", num_bytes, hostname.c_str(), mount_point.c_str());
-        auto host = simgrid::s4u::Host::by_name_or_null(hostname);
+
+        auto host = S4U_Simulation::get_host_or_vm_by_name_or_null(hostname);
         if (not host) {
             throw std::invalid_argument("S4U_Simulation::writeToDisk(): unknown host " + hostname);
         }
 
-        auto disk_list = simgrid::s4u::Host::by_name(hostname)->get_disks();
+        auto disk_list = host->get_disks();
         for (auto disk: disk_list) {
             std::string disk_mountpoint =
                     FileLocation::sanitizePath(std::string(std::string(disk->get_property("mount"))));
@@ -557,7 +561,7 @@ namespace wrench {
         simgrid::s4u::Disk *read_disk = nullptr;
         simgrid::s4u::Disk *write_disk = nullptr;
 
-        auto disk_list = simgrid::s4u::Host::by_name(hostname)->get_disks();
[…]
