+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -130,11 +130,15 @@ namespace wrench {
             WRENCH_DEBUG("Got a [%s] message", message->getName().c_str());
             if(reduced){//handle the 2 Advancecd Messages if reduced simulation is enabled for this node
                 if (auto msg = dynamic_cast<AdvancedContinueSearchMessage *>(message.get())) {
+                    WRENCH_DEBUG("Advanced Continue Search for %s",msg->file->getID().c_str());
+
                     S4U_Simulation::compute(this->getPropertyValueAsDouble(Property::CACHE_LOOKUP_OVERHEAD));
                     if(cached(msg->file)){//caching is even more efficient than advanced search
+                        WRENCH_DEBUG("File %s found in cache",msg->file->getID().c_str());
+
                         auto cached = getCached(msg->file);
                         try {
-                            S4U_Mailbox::putMessage(supervisor->mailbox,
+                            S4U_Mailbox::dputMessage(supervisor->mailbox,
                                                     new UpdateCacheMessage(
                                                             msg->answer_mailbox,
                                                             msg->original,
@@ -156,14 +160,16 @@ namespace wrench {
                             map<Node *, vector<stack<Node *>>> splitStacks = splitStack(msg->searchStack);
                             S4U_Simulation::compute(
                                     this->getPropertyValueAsDouble(Property::SEARCH_BROADCAST_OVERHEAD));
-                                for (auto entry: splitStacks) {
+                            WRENCH_DEBUG("Advanced Broadcast to %lu hosts",splitStacks.size());
+
+                            for (auto entry: splitStacks) {
                                     if (entry.first == this) {//this node was the target
                                         if (internalStorage &&//check the storage, it SHOULD be there, but we should check still
                                             StorageService::lookupFile(msg->file, FileLocation::LOCATION(internalStorage))) {
                                             //File in internal storage
                                             cache.add(msg->file, FileLocation::LOCATION(internalStorage));
                                             try {
-                                                S4U_Mailbox::putMessage(supervisor->mailbox,
+                                                S4U_Mailbox::dputMessage(supervisor->mailbox,
                                                                         new UpdateCacheMessage(
                                                                                 msg->answer_mailbox,
                                                                                 msg->original,
@@ -182,7 +188,7 @@ namespace wrench {
                                         }
                                     } else {
                                         try {
-                                            S4U_Mailbox::putMessage(entry.first->mailbox,
+                                            S4U_Mailbox::dputMessage(entry.first->mailbox,
                                                                     new AdvancedContinueSearchMessage(
                                                                             msg,
                                                                             entry.second
@@ -196,12 +202,13 @@ namespace wrench {
 
                                 }
                         }
+
                         if (internalStorage &&//check the storage, it SHOULD be there, but we should check still
                                     StorageService::lookupFile(msg->file, FileLocation::LOCATION(internalStorage))) {
                             //File in internal storage
                             cache.add(msg->file, FileLocation::LOCATION(internalStorage));
                             try {
-                                S4U_Mailbox::putMessage(supervisor->mailbox,
+                                S4U_Mailbox::dputMessage(supervisor->mailbox,
                                                         new UpdateCacheMessage(
                                                                 msg->answer_mailbox,
                                                                 msg->original,
@@ -243,7 +250,7 @@ namespace wrench {
                         for (auto entry: splitStacks) {
                             if (entry.first != this) {
                                 try {
[…]
