### 2022-07-20 00520d8
Added a no-op XRootD test
