### 2022-07-20 03c5eea
Added a no-op XRootD test
