### 2022-07-20 a410aa3
Added a no-op XRootD test
