+++ test/services/storage_services/XRootDStorageService/XRootDStorageServiceFunctionalTest.cpp
@@ -23,10 +23,13 @@ public:
 
     void do_BasicFunctionality_test();
 
+    std::shared_ptr<wrench::XRootD::Node> supervisor;
 
 protected:
 
     XRootDServiceFunctionalTest() {
+
+
 
         // Create a one-host platform file
         std::string xml = "<?xml version='1.0'?>"
@@ -65,6 +68,7 @@ protected:
     }
 
     std::string platform_file_path = UNIQUE_TMP_PATH_PREFIX + "platform.xml";
+
 };
 
 
@@ -76,7 +80,8 @@ class XRootDServiceBasicFunctionalityTestExecutionController : public wrench::Ex
 
 public:
     XRootDServiceBasicFunctionalityTestExecutionController(XRootDServiceFunctionalTest *test,
-                                                  std::string hostname) : wrench::ExecutionController(hostname, "test"), test(test) {
+                                                           std::string hostname) :
+            wrench::ExecutionController(hostname, "test"), test(test) {
     }
 
 private:
@@ -84,7 +89,28 @@ private:
 
     int main() {
 
-        WRENCH_INFO("I am a Controller");
+        WRENCH_INFO("Adding a file files to the simulation");
+        auto file1 = wrench::Simulation::addFile("file1", 10000);
+        auto file2 = wrench::Simulation::addFile("file2", 10000);
+
+        // TODO: The exception message generated by this sucks, we should fix it
+        // this->test->supervisor->createFile(file1);
+
+        // Create a copy file1 on the first child
+        wrench::Simulation::createFile(file1,
+        wrench::FileLocation::LOCATION(this->test->supervisor->getChild(0)));
+
+        // Read file1 from XRootD
+        this->test->supervisor->readFile(file1);
+
+        // Try to read file2 from XRootD although it is nowhere
+        // TODO: This should raise some FileNotFound Exception rather
+        // that propagating the network timeout. Come to think of it, one should probably
+        // set the search timeout as a special property, since the default 30-sec is WAY too long
+        // here.
+        this->test->supervisor->readFile(file2);
+
+
 
         return 0;
     }
@@ -111,7 +137,7 @@ void XRootDServiceFunctionalTest::do_BasicFunctionality_test() {
     // Create a XRootD Manager object
     wrench::XRootD::XRootD xrootdManager(simulation,{{wrench::XRootD::Property::CACHE_MAX_LIFETIME,"28800"},{wrench::XRootD::Property::REDUCED_SIMULATION,"true"}},{});
 
-    auto supervisor = xrootdManager.createSupervisor("Host1");
+    this->supervisor = xrootdManager.createSupervisor("Host1");
 
     auto ss2 = xrootdManager.createStorageServer("Host2", "/disk100", {}, {});
     auto ss3 = xrootdManager.createStorageServer("Host3", "/disk100", {}, {});
