+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -826,15 +826,7 @@ namespace wrench {
             return 0;
 
         }
-        /**
-        * @brief create a new file to the file server
-        * @param file: A shared pointer to the file to add
-        *
-        * @throw std::invalid_argument
-        */
-        void Node::createFile(const std::shared_ptr<DataFile> &file) {
-            metavisor->createFile(file,this->getSharedPtr<Node>());
-        }
+
         /**
         * @brief construct the path to all targets IF they are in the subtree
         * @param targets: All the nodes to search for
