+++ test/services/storage_services/XRootDStorageService/XRootDStorageServiceFunctionalTest.cpp
@@ -94,6 +94,7 @@ private:
         auto file2 = wrench::Simulation::addFile("file2", 10000);
 
         // TODO: The exception message generated by this sucks, we should fix it
+        // TODO: I changed it to be more informative
         // this->test->supervisor->createFile(file1);
 
         // Create a copy file1 on the first child
