### 2022-08-11 a22e371
Small fixes with internal exceptions coverage++
