+++ test/services/compute_services/bare_metal_compound_jobs/BareMetalComputeServiceOneActionTests.cpp
@@ -681,6 +681,7 @@ private:
         // Submit the job with bogus args
         std::vector<std::map<std::string, std::string>> bogus_args;
         bogus_args.push_back({{"bogus_action", "somehost"}});
+        bogus_args.push_back({{"my_computation", "somehost"}});
         bogus_args.push_back({{"my_computation", "Host4:30:123:12:ba"}});
         bogus_args.push_back({{"my_computation", "Host4:1"}});
         bogus_args.push_back({{"my_computation", "Host4:12"}});
