### 2022-08-11 4ca6335
Small fixes with internal exceptions coverage++
