+++ src/wrench/services/storage/StorageServiceMessage.cpp
@@ -17,6 +17,8 @@
 #include <wrench/services/compute/virtualized_cluster/VirtualizedClusterComputeService.h>
 #include <wrench/services/storage/StorageService.h>
 
+#include <utility>
+
 namespace wrench {
 
     /**
@@ -35,11 +37,14 @@ namespace wrench {
     */
     StorageServiceFreeSpaceRequestMessage::StorageServiceFreeSpaceRequestMessage(simgrid::s4u::Mailbox *answer_mailbox,
                                                                                  double payload)
-        : StorageServiceMessage(payload), answer_mailbox(answer_mailbox) {
+            : StorageServiceMessage(payload) {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (answer_mailbox == nullptr) {
             throw std::invalid_argument(
                     "StorageServiceFreeSpaceRequestMessage::StorageServiceFreeSpaceRequestMessage(): Invalid arguments");
         }
+#endif
+        this->answer_mailbox = answer_mailbox;
     }
 
     /**
@@ -51,14 +56,17 @@ namespace wrench {
      */
     StorageServiceFreeSpaceAnswerMessage::StorageServiceFreeSpaceAnswerMessage(
             std::map<std::string, double> free_space, double payload)
-        : StorageServiceMessage(payload) {
+            : StorageServiceMessage(payload) {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
+
         for (auto const &f: free_space) {
             if (f.second < 0) {
                 throw std::invalid_argument(
                         "StorageServiceFreeSpaceAnswerMessage::StorageServiceFreeSpaceAnswerMessage(): Invalid arguments");
             }
         }
-        this->free_space = free_space;
+#endif
+        this->free_space = std::move(free_space);
     }
 
     /**
@@ -74,11 +82,16 @@ namespace wrench {
                                                                                    std::shared_ptr<DataFile> file,
                                                                                    std::shared_ptr<FileLocation> location,
                                                                                    double payload)
-        : StorageServiceMessage(payload), answer_mailbox(answer_mailbox), file(file), location(location) {
+            : StorageServiceMessage(payload) {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if ((file == nullptr) || (location == nullptr) || (answer_mailbox == nullptr)) {
             throw std::invalid_argument(
                     "StorageServiceFileLookupRequestMessage::StorageServiceFileLookupRequestMessage(): Invalid arguments");
         }
+#endif
+        this->answer_mailbox = answer_mailbox;
+        this->file = std::move(file);
+        this->location = std::move(location);
     }
 
     /**
@@ -92,12 +105,14 @@ namespace wrench {
     StorageServiceFileLookupAnswerMessage::StorageServiceFileLookupAnswerMessage(std::shared_ptr<DataFile> file,
                                                                                  bool file_is_available,
                                                                                  double payload)
-        : StorageServiceMessage(payload) {
+            : StorageServiceMessage(payload) {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (file == nullptr) {
             throw std::invalid_argument(
                     "StorageServiceFileLookupAnswerMessage::StorageServiceFileLookupAnswerMessage(): Invalid arguments");
         }
-        this->file = file;
+#endif
+        this->file = std::move(file);
         this->file_is_available = file_is_available;
     }
 
@@ -114,11 +129,16 @@ namespace wrench {
                                                                                    std::shared_ptr<DataFile> file,
                                                                                    std::shared_ptr<FileLocation> location,
                                                                                    double payload)
[…]
