### 2022-08-11 b340df9
Small fixes with internal exceptions coverage++
