+++ src/wrench/managers/JobManagerMessage.cpp
@@ -30,6 +30,11 @@ namespace wrench {
                                                                                  std::shared_ptr<ComputeService> compute_service,
                                                                                  std::map<std::shared_ptr<WorkflowTask>, WorkflowTask::State> necessary_state_changes)
         : JobManagerMessage() {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
+        if ((job == nullptr) || (compute_service == nullptr)) {
+            throw std::invalid_argument("JobManagerStandardJobCompletedMessage::JobManagerStandardJobCompletedMessage(): Invalid arguments");
+        }
+#endif
         this->job = std::move(job);
         this->compute_service = std::move(compute_service);
         this->necessary_state_changes = std::move(necessary_state_changes);
@@ -48,6 +53,11 @@ namespace wrench {
                                                                            std::map<std::shared_ptr<WorkflowTask>, WorkflowTask::State> necessary_state_changes,
                                                                            std::set<std::shared_ptr<WorkflowTask>> necessary_failure_count_increments,
                                                                            std::shared_ptr<FailureCause> cause) : JobManagerMessage() {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
+        if ((job == nullptr) || (compute_service == nullptr) || (cause == nullptr)) {
+            throw std::invalid_argument("JobManagerStandardJobFailedMessage::JobManagerStandardJobFailedMessage(): Invalid arguments");
+        }
+#endif
         this->job = std::move(job);
         this->compute_service = std::move(compute_service);
         this->necessary_state_changes = std::move(necessary_state_changes);
@@ -64,6 +74,11 @@ namespace wrench {
     JobManagerPilotJobFailedMessage::JobManagerPilotJobFailedMessage(std::shared_ptr<PilotJob> job,
                                                                      std::shared_ptr<ComputeService> compute_service,
                                                                      std::shared_ptr<FailureCause> cause) : JobManagerMessage() {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
+        if ((job == nullptr) || (compute_service == nullptr)) {
+            throw std::invalid_argument("JobManagerPilotJobFailedMessage::JobManagerPilotJobFailedMessage(): Invalid arguments");
+        }
+#endif
         this->job = std::move(job);
         this->compute_service = std::move(compute_service);
         this->cause = std::move(cause);
@@ -77,6 +92,11 @@ namespace wrench {
      */
     JobManagerCompoundJobCompletedMessage::JobManagerCompoundJobCompletedMessage(std::shared_ptr<CompoundJob> job,
                                                                                  std::shared_ptr<ComputeService> compute_service) : JobManagerMessage() {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
+        if ((job == nullptr) || (compute_service == nullptr)) {
+            throw std::invalid_argument("JobManagerCompoundJobCompletedMessage::JobManagerCompoundJobCompletedMessage(): Invalid arguments");
+        }
+#endif
         this->job = std::move(job);
         this->compute_service = std::move(compute_service);
     }
@@ -90,6 +110,11 @@ namespace wrench {
     JobManagerCompoundJobFailedMessage::JobManagerCompoundJobFailedMessage(std::shared_ptr<CompoundJob> job,
                                                                            std::shared_ptr<ComputeService> compute_service,
                                                                            std::shared_ptr<FailureCause> cause) : JobManagerMessage() {
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
+        if ((job == nullptr) || (compute_service == nullptr)) {
[…]
