+++ src/wrench/services/compute/ComputeServiceMessage.cpp
@@ -34,16 +34,17 @@ namespace wrench {
             simgrid::s4u::Mailbox *answer_mailbox,
             std::shared_ptr<CompoundJob> job,
             std::map<std::string, std::string> service_specific_args,
-            double payload) : ComputeServiceMessage(payload),
-                              answer_mailbox(answer_mailbox), job(std::move(job)), service_specific_args(std::move(service_specific_args)) {
+            double payload) : ComputeServiceMessage(payload) {
 #ifdef WRENCH_INTERNAL_EXCEPTIONS
         if ((answer_mailbox == nullptr) || (job == nullptr)) {
             throw std::invalid_argument(
                     "ComputeServiceSubmitCompoundJobRequestMessage::ComputeServiceSubmitCompoundJobRequestMessage(): Invalid arguments");
         }
-    }
 #endif
-}
+        this->answer_mailbox = answer_mailbox;
+        this->job = std::move(job);
+        this->service_specific_args = std::move(service_specific_args);
+    }
 
 /**
      * @brief Constructor
@@ -55,24 +56,24 @@ namespace wrench {
      *
      * @throw std::invalid_arguments
      */
-ComputeServiceSubmitCompoundJobAnswerMessage::ComputeServiceSubmitCompoundJobAnswerMessage(std::shared_ptr<CompoundJob> job,
-                                                                                           std::shared_ptr<ComputeService> compute_service,
-                                                                                           bool success,
-                                                                                           std::shared_ptr<FailureCause> failure_cause,
-                                                                                           double payload) : ComputeServiceMessage(payload) {
+    ComputeServiceSubmitCompoundJobAnswerMessage::ComputeServiceSubmitCompoundJobAnswerMessage(std::shared_ptr<CompoundJob> job,
+                                                                                               std::shared_ptr<ComputeService> compute_service,
+                                                                                               bool success,
+                                                                                               std::shared_ptr<FailureCause> failure_cause,
+                                                                                               double payload) : ComputeServiceMessage(payload) {
 #ifdef WRENCH_INTERNAL_EXCEPTIONS
-    if ((job == nullptr) || (compute_service == nullptr) ||
-        (success && (failure_cause != nullptr)) ||
-        (!success && (failure_cause == nullptr))) {
-        throw std::invalid_argument(
-                "ComputeServiceSubmitCompoundJobAnswerMessage::ComputeServiceSubmitCompoundJobAnswerMessage(): Invalid arguments");
-    }
+        if ((job == nullptr) || (compute_service == nullptr) ||
+            (success && (failure_cause != nullptr)) ||
+            (!success && (failure_cause == nullptr))) {
+            throw std::invalid_argument(
+                    "ComputeServiceSubmitCompoundJobAnswerMessage::ComputeServiceSubmitCompoundJobAnswerMessage(): Invalid arguments");
+        }
 #endif
-    this->job = std::move(job);
-    this->compute_service = std::move(compute_service);
-    this->success = success;
-    this->failure_cause = std::move(failure_cause);
-}
+        this->job = std::move(job);
+        this->compute_service = std::move(compute_service);
+        this->success = success;
+        this->failure_cause = std::move(failure_cause);
+    }
 
 /**
      * @brief Constructor
@@ -82,20 +83,20 @@ ComputeServiceSubmitCompoundJobAnswerMessage::ComputeServiceSubmitCompoundJobAns
      *
      * @throw std::invalid_arguments
      */
[…]
