### 2022-08-11 603bc93
Committing clang-format changes
