+++ src/wrench/services/compute/ComputeServiceMessage.cpp
@@ -46,7 +46,7 @@ namespace wrench {
         this->service_specific_args = std::move(service_specific_args);
     }
 
-/**
+    /**
      * @brief Constructor
      * @param job: a compound job that had been submitted for execution
      * @param compute_service: the compute service
@@ -75,7 +75,7 @@ namespace wrench {
         this->failure_cause = std::move(failure_cause);
     }
 
-/**
+    /**
      * @brief Constructor
      * @param job: a compound job that has completed
      * @param cs: the compute service on which the job has completed
@@ -86,7 +86,7 @@ namespace wrench {
     ComputeServiceCompoundJobDoneMessage::ComputeServiceCompoundJobDoneMessage(std::shared_ptr<CompoundJob> job,
                                                                                std::shared_ptr<ComputeService> cs,
                                                                                double payload)
-            : ComputeServiceMessage(payload) {
+        : ComputeServiceMessage(payload) {
 
 #ifdef WRENCH_INTERNAL_EXCEPTIONS
         if ((job == nullptr) || (cs == nullptr)) {
@@ -98,7 +98,7 @@ namespace wrench {
         this->compute_service = std::move(cs);
     }
 
-/**
+    /**
      * @brief Constructor
      * @param job: a compound job that has failed
      * @param cs: the compute service on which the job has failed
@@ -109,7 +109,7 @@ namespace wrench {
     ComputeServiceCompoundJobFailedMessage::ComputeServiceCompoundJobFailedMessage(std::shared_ptr<CompoundJob> job,
                                                                                    std::shared_ptr<ComputeService> cs,
                                                                                    double payload)
-            : ComputeServiceMessage(payload) {
+        : ComputeServiceMessage(payload) {
 #ifdef WRENCH_INTERNAL_EXCEPTIONS
         if ((job == nullptr) || (cs == nullptr)) {
             throw std::invalid_argument(
@@ -120,7 +120,7 @@ namespace wrench {
         this->compute_service = std::move(cs);
     }
 
-/**
+    /**
     * @brief Constructor
     * @param answer_mailbox: mailbox to which the answer message should be sent
     * @param job: a compound job to terminate
@@ -142,7 +142,7 @@ namespace wrench {
         this->job = std::move(job);
     }
 
-/**
+    /**
      * @brief Constructor
      * @param job: a compound job whose termination was requested
      * @param compute_service: the compute service that was executing the standard job
@@ -171,60 +171,60 @@ namespace wrench {
         this->failure_cause = std::move(failure_cause);
     }
 
-//    /**
-//     * @brief Constructor
-//     * @param answer_mailbox: mailbox to which the answer message should be sent
-//     * @param job: a pilot job submitted for execution
-//     * @param service_specific_args: a map of extra arguments (each specified by a name and value, both strings) required by some services
-//     * @param payload: message size in bytes
-//     *
-//     * @throw std::invalid_argument
-//     */
-//    ComputeServiceSubmitPilotJobRequestMessage::ComputeServiceSubmitPilotJobRequestMessage(std::string answer_mailbox,
-//                                                                                           std::shared_ptr<PilotJob> job,
-//                                                                                           const std::map<std::string, std::string> service_specific_args,
-//                                                                                           double payload)
-//            : ComputeServiceMessage(
-//            "SUBMIT_PILOT_JOB_REQUEST", payload) {
-//        if ((job == nullptr) || (answer_mailbox == "")) {
-//            throw std::invalid_argument(
-//                    "ComputeServiceSubmitPilotJobRequestMessage::ComputeServiceSubmitPilotJobRequestMessage(): Invalid arguments");
-//        }
-//        this->answer_mailbox = answer_mailbox;
[…]
