### 2022-08-11 573d05e
Committing clang-format changes
