### 2022-08-11 f493a3e
Committing clang-format changes
