+++ test/simulation/SimpleSimulationTest.cpp
@@ -166,6 +166,9 @@ private:
         // Get pointer to cloud service
         auto cs = this->test->compute_service;
 
+        // Coverage
+        cs->getPhysicalHostname();
+
         // Create a bogus VM
         try {
             cs->createVM(1000, 10000000);
@@ -329,16 +332,16 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
     // Adding services to an uninitialized simulation
     std::vector<std::string> hosts = {"DualCoreHost", "QuadCoreHost"};
     ASSERT_THROW(simulation->add(
-                         new wrench::CloudComputeService("DualCoreHost", hosts, "/scratch")),
+            new wrench::CloudComputeService("DualCoreHost", hosts, "/scratch")),
                  std::runtime_error);
     ASSERT_THROW(simulation->add(
-                         new wrench::SimpleStorageService("DualCoreHost", {"/"})),
+            new wrench::SimpleStorageService("DualCoreHost", {"/"})),
                  std::runtime_error);
     ASSERT_THROW(simulation->add(
-                         new wrench::NetworkProximityService("DualCoreHost", hosts)),
+            new wrench::NetworkProximityService("DualCoreHost", hosts)),
                  std::runtime_error);
     ASSERT_THROW(simulation->add(
-                         new wrench::FileRegistryService("DualCoreHost")),
+            new wrench::FileRegistryService("DualCoreHost")),
                  std::runtime_error);
 
     ASSERT_NO_THROW(simulation->init(&argc, argv));
@@ -352,16 +355,22 @@ void SimpleSimulationTest::do_getReadyTasksTest_test() {
 
     // Create a Storage Service
     ASSERT_THROW(storage_service = simulation->add(
-                         new wrench::SimpleStorageService(hostname, {"/disk1"}, {},
-                                                          {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, -1}})),
+            new wrench::SimpleStorageService(hostname, {"/disk1"}, {},
+                                             {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, -1}})),
                  std::invalid_argument);
 
     storage_service = simulation->add(
             new wrench::SimpleStorageService(hostname, {"/disk2"},
-                                             {{wrench::SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS, "567"},
-                                              {wrench::ServiceProperty::translateString("StorageServiceProperty::BUFFER_SIZE"), "678"}},
+                                             {
+                                                     {wrench::SimpleStorageServiceProperty::MAX_NUM_CONCURRENT_DATA_CONNECTIONS, "567"},
+                                                     {wrench::ServiceProperty::translateString("StorageServiceProperty::BUFFER_SIZE"), "678"}
+                                             },
                                              {{wrench::SimpleStorageServiceMessagePayload::FILE_COPY_ANSWER_MESSAGE_PAYLOAD, 123},
                                               {wrench::ServiceMessagePayload::translateString("StorageServiceMessagePayload::FILE_LOOKUP_ANSWER_MESSAGE_PAYLOAD"), 234}}));
+
+    // Coverage
+    ASSERT_GE(storage_service->getPropertyList().size(), 2);
[…]
