### 2022-08-11 7f7c51b
Committing clang-format changes
