+++ test/services/helper_services/action_executor/SleepActionExecutorTest.cpp
@@ -129,6 +129,10 @@ private:
         auto job = job_manager->createCompoundJob("");
         // Add a sleep_action
         auto sleep_action = job->addSleepAction("", 10.0);
+
+        // coverage
+        sleep_action->getActionTypeAsString();
+
         // Create a sleep action executor
         auto sleep_action_executor = std::shared_ptr<wrench::ActionExecutor>(
                 new wrench::ActionExecutor("Host2", 0, 0.0, 0, false, this->mailbox, sleep_action, nullptr));
