+++ test/services/helper_services/action_executor/CustomActionExecutorTest.cpp
@@ -150,6 +150,9 @@ private:
 
         action = std::dynamic_pointer_cast<wrench::Action>(job->addCustomAction("", 0, 0, lambda_execute, lambda_terminate));
 
+        // coverage
+        action->getActionTypeAsString();
+
         auto action_executor = std::shared_ptr<wrench::ActionExecutor>(
                 new wrench::ActionExecutor("Host2",
                                            num_cores,
