+++ test/services/helper_services/action_executor/FileReadActionExecutorTest.cpp
@@ -139,9 +139,13 @@ private:
         // Add a file_read_action
         auto file_read_action = job->addFileReadAction("", this->test->file,
                                                        wrench::FileLocation::LOCATION(this->test->ss));
+        // coverage
+        file_read_action->getActionTypeAsString();
+
         // Create a file read action executor
         auto file_read_action_executor = std::shared_ptr<wrench::ActionExecutor>(
                 new wrench::ActionExecutor("Host2", 0, 0.0, 0, false, this->mailbox, file_read_action, nullptr));
+
         // Start it
         file_read_action_executor->setSimulation(this->simulation);
         file_read_action_executor->start(file_read_action_executor, true, false);
