+++ test/services/helper_services/action_executor/FileDeleteActionExecutorTest.cpp
@@ -132,6 +132,10 @@ private:
         // Add a file_delete_action
         auto file_delete_action = job->addFileDeleteAction("", this->test->file,
                                                            wrench::FileLocation::LOCATION(this->test->ss));
+
+        // coverage
+        file_delete_action->getActionTypeAsString();
+
         // Create a file read action executor
         auto file_delete_action_executor = std::shared_ptr<wrench::ActionExecutor>(
                 new wrench::ActionExecutor("Host2", 0, 0.0, 0, false, this->mailbox, file_delete_action, nullptr));
