+++ test/services/helper_services/action_executor/ComputeActionExecutorTest.cpp
@@ -143,7 +143,7 @@ private:
         action = std::dynamic_pointer_cast<wrench::Action>(job->addComputeAction("", 20.0, 100.0, 1, 4, wrench::ParallelModel::AMDAHL(1.0)));
 
         // coverage
-        action->getActionTypeAsString();
+        wrench::Action::getActionTypeAsString(action);
 
         auto action_executor = std::shared_ptr<wrench::ActionExecutor>(
                 new wrench::ActionExecutor("Host2",
