+++ test/services/helper_services/action_executor/FileReadActionExecutorTest.cpp
@@ -140,7 +140,7 @@ private:
         auto file_read_action = job->addFileReadAction("", this->test->file,
                                                        wrench::FileLocation::LOCATION(this->test->ss));
         // coverage
-        file_read_action->getActionTypeAsString();
+        wrench::Action::getActionTypeAsString(file_read_action);
 
         // Create a file read action executor
         auto file_read_action_executor = std::shared_ptr<wrench::ActionExecutor>(
