### 2022-08-12 87042d1
Committing clang-format changes
