### 2022-08-13 d6af6ea
Committing clang-format changes
