### 2022-08-13 481579a
Committing clang-format changes
