### 2022-08-13 6812e2d
Committing clang-format changes
