+++ test/services/compute_services/bare_metal_standard_jobs/BareMetalComputeServiceOneTaskTest.cpp
@@ -202,18 +202,18 @@ void BareMetalComputeServiceOneTaskTest::do_BadSetup_test() {
 
     // Empty resource list
     ASSERT_THROW(compute_service = simulation->add(
-            new wrench::BareMetalComputeService("bogus",
-                                                (std::map<std::string, std::tuple<unsigned long, double>>){},
-                                                {})),
+                         new wrench::BareMetalComputeService("bogus",
+                                                             (std::map<std::string, std::tuple<unsigned long, double>>){},
+                                                             {})),
                  std::invalid_argument);
 
     // Bad hostname
     ASSERT_THROW(compute_service = simulation->add(
-            new wrench::BareMetalComputeService("bogus",
-                                                {std::make_pair("bogus",
-                                                                std::make_tuple(wrench::ComputeService::ALL_CORES,
-                                                                                wrench::ComputeService::ALL_RAM))},
-                                                {})),
+                         new wrench::BareMetalComputeService("bogus",
+                                                             {std::make_pair("bogus",
+                                                                             std::make_tuple(wrench::ComputeService::ALL_CORES,
+                                                                                             wrench::ComputeService::ALL_RAM))},
+                                                             {})),
                  std::invalid_argument);
 
     // Get a hostname
@@ -221,63 +221,63 @@ void BareMetalComputeServiceOneTaskTest::do_BadSetup_test() {
 
     // Bad resource hostname
     ASSERT_THROW(compute_service = simulation->add(
-            new wrench::BareMetalComputeService(hostname,
-                                                {std::make_pair("bogus",
-                                                                std::make_tuple(wrench::ComputeService::ALL_CORES,
-                                                                                wrench::ComputeService::ALL_RAM))},
-                                                "",
-                                                {})),
+                         new wrench::BareMetalComputeService(hostname,
+                                                             {std::make_pair("bogus",
+                                                                             std::make_tuple(wrench::ComputeService::ALL_CORES,
+                                                                                             wrench::ComputeService::ALL_RAM))},
+                                                             "",
+                                                             {})),
                  std::invalid_argument);
 
     // Bad number of cores
     ASSERT_THROW(compute_service = simulation->add(
-            new wrench::BareMetalComputeService(hostname,
-                                                {std::make_pair(hostname,
-                                                                std::make_tuple(0,
-                                                                                wrench::ComputeService::ALL_RAM))},
-                                                "",
-                                                {})),
+                         new wrench::BareMetalComputeService(hostname,
+                                                             {std::make_pair(hostname,
+                                                                             std::make_tuple(0,
+                                                                                             wrench::ComputeService::ALL_RAM))},
+                                                             "",
+                                                             {})),
                  std::invalid_argument);
 
     // Bad number of cores
     ASSERT_THROW(compute_service = simulation->add(
-            new wrench::BareMetalComputeService(hostname,
-                                                {std::make_pair(hostname,
-                                                                std::make_tuple(100,
-                                                                                wrench::ComputeService::ALL_RAM))},
-                                                {})),
+                         new wrench::BareMetalComputeService(hostname,
+                                                             {std::make_pair(hostname,
+                                                                             std::make_tuple(100,
+                                                                                             wrench::ComputeService::ALL_RAM))},
+                                                             {})),
                  std::invalid_argument);
 
 
     // Bad RAM
     ASSERT_THROW(compute_service = simulation->add(
-            new wrench::BareMetalComputeService(hostname,
-                                                {std::make_pair("RAMHost",
-                                                                std::make_tuple(wrench::ComputeService::ALL_CORES,
-                                                                                -1.0))},
[…]
