### 2022-08-13 481ff18
Committing clang-format changes
