### 2022-08-14 027b8ce
Committing clang-format changes
