+++ test/wms/WMSTest.cpp
@@ -359,8 +359,6 @@ private:
         }
 
 
-
-
         return 0;
     }
 
