+++ test/services/helper_services/action_executor/ComputeActionExecutorTest.cpp
@@ -37,7 +37,8 @@ public:
     std::shared_ptr<wrench::DataFile> file;
     std::shared_ptr<wrench::StorageService> ss;
 
-    void do_ComputeActionExecutorSuccessTest_test();
+    void do_ComputeActionExecutorSuccessTest_test(bool simulation_computation_as_sleep);
+    void do_ComputeActionExecutorFailureTest_test();
 
 protected:
     ComputeActionExecutorTest() {
@@ -120,12 +121,14 @@ class ComputeActionExecutorTestWMS : public wrench::ExecutionController {
 
 public:
     ComputeActionExecutorTestWMS(ComputeActionExecutorTest *test,
-                                 std::string hostname) : wrench::ExecutionController(hostname, "test"), test(test) {
+                                 std::string hostname,
+                                 bool simulation_computation_as_sleep) : wrench::ExecutionController(hostname, "test"), test(test), simulation_computation_as_sleep(simulation_computation_as_sleep) {
     }
 
 
 private:
     ComputeActionExecutorTest *test;
+    bool simulation_computation_as_sleep;
 
     int main() {
 
@@ -140,7 +143,8 @@ private:
         unsigned long num_cores = 2;
         double ram = 200;
 
-        action = std::dynamic_pointer_cast<wrench::Action>(job->addComputeAction("", 20.0, 100.0, 1, 4, wrench::ParallelModel::AMDAHL(1.0)));
+        action = std::dynamic_pointer_cast<wrench::Action>(
+                job->addComputeAction("", 20.0, 100.0, 1, 4, wrench::ParallelModel::AMDAHL(1.0)));
 
         // coverage
         wrench::Action::getActionTypeAsString(action);
@@ -150,7 +154,7 @@ private:
                                            num_cores,
                                            ram,
                                            0.1,
-                                           false,
+                                           simulation_computation_as_sleep,
                                            this->mailbox,
                                            action,
                                            nullptr));
@@ -193,10 +197,11 @@ private:
 };
 
 TEST_F(ComputeActionExecutorTest, Success) {
-    DO_TEST_WITH_FORK(do_ComputeActionExecutorSuccessTest_test);
+    DO_TEST_WITH_FORK_ONE_ARG(do_ComputeActionExecutorSuccessTest_test, true);
+    DO_TEST_WITH_FORK_ONE_ARG(do_ComputeActionExecutorSuccessTest_test, false);
 }
 
-void ComputeActionExecutorTest::do_ComputeActionExecutorSuccessTest_test() {
+void ComputeActionExecutorTest::do_ComputeActionExecutorSuccessTest_test(bool simulation_computation_as_sleep) {
 
     // Create and initialize a simulation
     simulation = wrench::Simulation::createSimulation();
@@ -224,7 +229,130 @@ void ComputeActionExecutorTest::do_ComputeActionExecutorSuccessTest_test() {
     // Create a WMS
     std::shared_ptr<wrench::ExecutionController> wms = nullptr;
     ASSERT_NO_THROW(wms = simulation->add(
-                            new ComputeActionExecutorTestWMS(this, "Host1")));
+                            new ComputeActionExecutorTestWMS(this, "Host1", simulation_computation_as_sleep)));
+
+    ASSERT_NO_THROW(simulation->launch());
+
+    this->workflow->clear();
+
+    for (int i = 0; i < argc; i++)
+        free(argv[i]);
+    free(argv);
+}
+
+
+
+/**********************************************************************/
+/**  COMPUTE ACTION EXECUTOR FAILURE TEST                            **/
+/**********************************************************************/
+
+
+class ComputeActionExecutorFailureTestWMS : public wrench::ExecutionController {
+
+public:
[…]
