+++ test/services/helper_services/action_executor/FileRegistryActionExecutorTest.cpp
@@ -10,7 +10,7 @@
 #include <gtest/gtest.h>
 #include <wrench-dev.h>
 
-#include <wrench/action/FileDeleteAction.h>
+#include <wrench/action/FileRegistryAction.h>
 #include <wrench/services/helper_services/action_executor/ActionExecutorMessage.h>
 #include <wrench/services/helper_services/action_executor/ActionExecutor.h>
 #include <wrench/job/CompoundJob.h>
@@ -21,20 +21,17 @@
 #include "../../../include/TestWithFork.h"
 #include "../../../include/UniqueTmpPathPrefix.h"
 
-WRENCH_LOG_CATEGORY(file_delete_action_executor_test, "Log category for FileDeleteActionExecutorTest");
+WRENCH_LOG_CATEGORY(file_registry_action_executor_test, "Log category for FileRegistryActionExecutorTest");
 
-#define EPSILON (std::numeric_limits<double>::epsilon())
-
-class FileDeleteActionExecutorTest : public ::testing::Test {
+class FileRegistryActionExecutorTest : public ::testing::Test {
 
 public:
     std::shared_ptr<wrench::Simulation> simulation;
 
-    void do_FileDeleteActionExecutorSuccessTest_test();
-
+    void do_FileRegistryActionExecutorSuccessTest_test();
 
 protected:
-    FileDeleteActionExecutorTest() {
+    FileRegistryActionExecutorTest() {
 
         std::string xml = "<?xml version='1.0'?>"
                           "<!DOCTYPE platform SYSTEM \"https://simgrid.org/simgrid.dtd\">"
@@ -101,91 +98,130 @@ protected:
     }
 
     std::string platform_file_path = UNIQUE_TMP_PATH_PREFIX + "platform.xml";
+    std::shared_ptr<wrench::Workflow> workflow;
 
 public:
     std::shared_ptr<wrench::DataFile> file;
     std::shared_ptr<wrench::StorageService> ss;
+    std::shared_ptr<wrench::FileRegistryService> fr;
 };
 
 
 /**********************************************************************/
-/**  DO FILE_READ ACTION EXECUTOR SUCCESS TEST                       **/
+/**  DO FILE_REGISTRY ACTION EXECUTOR SUCCESS TEST                   **/
 /**********************************************************************/
 
-class FileDeleteActionExecutorSuccessTestWMS : public wrench::ExecutionController {
+class FileRegistryActionExecutorSuccessTestWMS : public wrench::ExecutionController {
 
 public:
-    FileDeleteActionExecutorSuccessTestWMS(FileDeleteActionExecutorTest *test,
-                                           std::string hostname) : wrench::ExecutionController(hostname, "test"), test(test) {
+    FileRegistryActionExecutorSuccessTestWMS(FileRegistryActionExecutorTest *test,
+                                         std::shared_ptr<wrench::Workflow> workflow,
+                                         std::string hostname) : wrench::ExecutionController(hostname, "test"), test(test), workflow(workflow) {
     }
 
 private:
-    FileDeleteActionExecutorTest *test;
+    FileRegistryActionExecutorTest *test;
+    std::shared_ptr<wrench::Workflow> workflow;
 
     int main() {
 
         // Create a job manager
         auto job_manager = this->createJobManager();
 
-        // Create a compound job
-        auto job = job_manager->createCompoundJob("");
-        // Add a file_delete_action
-        auto file_delete_action = job->addFileDeleteAction("", this->test->file,
-                                                           wrench::FileLocation::LOCATION(this->test->ss));
+        {
+            // Create a compound job
+            auto job = job_manager->createCompoundJob("");
+            // Add a file_read_action
+            auto file_registry_add_entry_action = job->addFileRegistryAddEntryAction("", this->test->fr,
[…]
