### 2022-08-16 b2043fc
Committing clang-format changes
