### 2022-08-17 0edfd45
Committing clang-format changes
