+++ test/workflow/WorkflowTest.cpp
@@ -390,36 +390,36 @@ TEST_F(AllDependenciesWorkflowTest, AllDependenciesWorkflowStructure) {
 TEST_F(WorkflowTest, LowLevelDagOfTasksTest) {
     wrench::DagOfTasks dag;
 
-    ASSERT_NO_THROW(dag.addVertex((wrench::WorkflowTask *)1));
-    ASSERT_NO_THROW(dag.removeVertex((wrench::WorkflowTask *)1));
-    ASSERT_THROW(dag.removeVertex((wrench::WorkflowTask *)1), std::invalid_argument);
-    ASSERT_NO_THROW(dag.addVertex((wrench::WorkflowTask *)1));
-    ASSERT_NO_THROW(dag.addVertex((wrench::WorkflowTask *)2));
+    ASSERT_NO_THROW(dag.addVertex((wrench::WorkflowTask *) 1));
+    ASSERT_NO_THROW(dag.removeVertex((wrench::WorkflowTask *) 1));
+    ASSERT_THROW(dag.removeVertex((wrench::WorkflowTask *) 1), std::invalid_argument);
+    ASSERT_NO_THROW(dag.addVertex((wrench::WorkflowTask *) 1));
+    ASSERT_NO_THROW(dag.addVertex((wrench::WorkflowTask *) 2));
 
-    ASSERT_NO_THROW(dag.addEdge((wrench::WorkflowTask *)1, (wrench::WorkflowTask *)2));
-    ASSERT_NO_THROW(dag.removeEdge((wrench::WorkflowTask *)1, (wrench::WorkflowTask *)2));
-    ASSERT_THROW(dag.removeEdge((wrench::WorkflowTask *)1, (wrench::WorkflowTask *)3), std::invalid_argument);
-    ASSERT_THROW(dag.removeEdge((wrench::WorkflowTask *)3, (wrench::WorkflowTask *)1), std::invalid_argument);
-    ASSERT_THROW(dag.addEdge((wrench::WorkflowTask *)1, (wrench::WorkflowTask *)3), std::invalid_argument);
-    ASSERT_THROW(dag.addEdge((wrench::WorkflowTask *)3, (wrench::WorkflowTask *)1), std::invalid_argument);
+    ASSERT_NO_THROW(dag.addEdge((wrench::WorkflowTask *) 1, (wrench::WorkflowTask *) 2));
+    ASSERT_NO_THROW(dag.removeEdge((wrench::WorkflowTask *) 1, (wrench::WorkflowTask *) 2));
+    ASSERT_THROW(dag.removeEdge((wrench::WorkflowTask *) 1, (wrench::WorkflowTask *) 3), std::invalid_argument);
+    ASSERT_THROW(dag.removeEdge((wrench::WorkflowTask *) 3, (wrench::WorkflowTask *) 1), std::invalid_argument);
+    ASSERT_THROW(dag.addEdge((wrench::WorkflowTask *) 1, (wrench::WorkflowTask *) 3), std::invalid_argument);
+    ASSERT_THROW(dag.addEdge((wrench::WorkflowTask *) 3, (wrench::WorkflowTask *) 1), std::invalid_argument);
 
-    ASSERT_NO_THROW(dag.doesPathExist((wrench::WorkflowTask *)1, (wrench::WorkflowTask *)2));
-    ASSERT_THROW(dag.doesPathExist((wrench::WorkflowTask *)1, (wrench::WorkflowTask *)3), std::invalid_argument);
-    ASSERT_THROW(dag.doesPathExist((wrench::WorkflowTask *)3, (wrench::WorkflowTask *)1), std::invalid_argument);
[…]
