### 2022-08-17 c190dbc
Make exceptions internal
