+++ src/wrench/simulation/SimulationTimestampTypes.cpp
@@ -60,10 +60,12 @@ namespace wrench {
      */
     SimulationTimestampTask::SimulationTimestampTask(double date, const std::shared_ptr<WorkflowTask> &task) : task(task) {
         this->date = date;
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (task == nullptr) {
             throw std::invalid_argument(
                     "SimulationTimestampTask::SimulationTimestampTask() requires a pointer to a Workflowtask");
         }
+#endif
     }
 
     /**
@@ -120,10 +122,12 @@ namespace wrench {
     SimulationTimestampTaskStart::SimulationTimestampTaskStart(double date, const std::shared_ptr<WorkflowTask> &task) : SimulationTimestampTask(date, task) {
         WRENCH_DEBUG("Inserting a Taskstart timestamp for task '%s'", task->getID().c_str());
 
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (task == nullptr) {
             throw std::invalid_argument(
                     "SimulationTimestampTaskStart::SimulationTimestampTaskStart() requires a valid pointer to a WorkflowTask object");
         }
+#endif
 
         /*
          * Upon creation, this object adds a pointer of itself to the 'pending_task_timestamps' map so that it's endpoint can
@@ -141,10 +145,12 @@ namespace wrench {
     SimulationTimestampTaskFailure::SimulationTimestampTaskFailure(double date, const std::shared_ptr<WorkflowTask> &task) : SimulationTimestampTask(date, task) {
         WRENCH_DEBUG("Inserting a TaskFailure timestamp for task '%s'", task->getID().c_str());
 
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (task == nullptr) {
             throw std::invalid_argument(
                     "SimulationTimestampTaskFailure::SimulationTimestampTaskFailure() requires a valid pointer to a WorkflowTask object");
         }
+#endif
 
         // match this timestamp with a SimulationTimestampTaskStart
         setEndpoints();
@@ -158,10 +164,12 @@ namespace wrench {
     SimulationTimestampTaskCompletion::SimulationTimestampTaskCompletion(double date, const std::shared_ptr<WorkflowTask> &task) : SimulationTimestampTask(date, task) {
         WRENCH_DEBUG("Inserting a TaskCompletion timestamp for task '%s'", task->getID().c_str());
 
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (task == nullptr) {
             throw std::invalid_argument(
                     "SimulationTimestampTaskCompletion::SimulationTimestampTaskCompletion() requires a valid pointer to a WorkflowTask object");
         }
+#endif
 
         // match this timestamp with a SimulationTimestampTaskStart
         setEndpoints();
@@ -176,10 +184,12 @@ namespace wrench {
         : SimulationTimestampTask(date, task) {
         WRENCH_DEBUG("Inserting a TaskTerminated timestamp for task '%s'", task->getID().c_str());
 
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
         if (task == nullptr) {
             throw std::invalid_argument(
                     "SimulationTimestampTaskTerminated::SimulationTimestampTaskTerminated() requires a valid pointer to a WorkflowTask object");
         }
+#endif
 
         // match this timestamp with a SimulationTimestampTaskStart
         setEndpoints();
@@ -271,10 +281,12 @@ namespace wrench {
         WRENCH_DEBUG("Inserting a FileCopyStart timestamp for file copy");
 
         // all information about a file copy should be passed
+#ifdef WRENCH_INTERNAL_EXCEPTIONS
[…]
