### 2022-08-24 3faa18e
Committing clang-format changes
