### 2022-08-24 b054f42
Committing clang-format changes
