### 2022-08-26 1e4a65a
StorageService now is incharge of createFile
