+++ include/wrench/services/storage/xrootd/Node.h
@@ -74,7 +74,10 @@ namespace wrench {
 
             bool cached(shared_ptr<DataFile> file);
             std::set<std::shared_ptr<FileLocation>> getCached(shared_ptr<DataFile> file);
-            void createFile(const std::shared_ptr<DataFile> &file)   ;
+            void createFile(const std::shared_ptr<DataFile> &file);
+
+            void createFile(const std::shared_ptr<DataFile> &file, const string & path);
+            void createFile(const std::shared_ptr<DataFile> &file,  const std::shared_ptr<FileLocation> &location);
             double getLoad() override;
             /***********************/
             /** \cond INTERNAL     */
