### 2022-08-26 52306a6
StorageService now is incharge of createFile
