+++ src/wrench/simulation/Simulation.cpp
@@ -740,7 +740,6 @@ namespace wrench {
             frs->addEntryToDatabase(file, location);
         }
     }
-
     /**
      * @brief Store a file at a particular mount point ex-nihilo. Doesn't notify a file registry service and will do nothing (and won't complain) if the file already exists
      * at that location.
@@ -750,9 +749,11 @@ namespace wrench {
      *
      * @throw std::invalid_argument
      */
+
     void Simulation::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<FileLocation> &location) {
-        location->getStorageService()->stageFile(file, location->getMountPoint(),
-                                                 location->getAbsolutePathAtMountPoint());
+        //location->getStorageService()->stageFile(file, location->getMountPoint(),
+        //                                         location->getAbsolutePathAtMountPoint());
+        location->getStorageService()->createFile(file,location);
     }
     /**
      * @brief Store a file on a particular file server ex-nihilo. Doesn't notify a file registry service and will do nothing (and won't complain) if the file already exists
@@ -762,8 +763,9 @@ namespace wrench {
      *
      * @throw std::invalid_argument
      */
+    [[deprecated("Replaced by StorageService::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<FileLocation> &location), do not use if using XRootD or other distributed file system")]]
     void Simulation::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<StorageService> &server) {
-        createFile(file, FileLocation::LOCATION(server));
+        server->createFile(file);
     }
 
     /**
