### 2022-08-26 3c22173
StorageService now is incharge of createFile
