+++ test/simulation/simulation_output/SimulationTimestampFileWriteTest.cpp
@@ -145,10 +145,10 @@ TEST_F(SimulationTimestampFileWriteTest, SimulationTimestampFileWriteBasicTest) 
 
 void SimulationTimestampFileWriteTest::do_SimulationTimestampFileWriteBasic_test() {
     auto simulation = wrench::Simulation::createSimulation();
-    int argc = 2;
+    int argc = 1;
     auto argv = (char **) calloc(argc, sizeof(char *));
     argv[0] = strdup("unit_test");
-    argv[1] = strdup("--wrench-full-log");
+//    argv[1] = strdup("--wrench-full-log");
 
     ASSERT_NO_THROW(simulation->init(&argc, argv));
 
