### 2022-08-26 c39bee0
Committing clang-format changes
