+++ examples/action_api/XRootD/stress-test/src/platformCreator.cpp
@@ -1,0 +1,269 @@
+#include "../include/platformCreator.h"
+#include <math.h>
+#include <vector>
+#include <cstdlib>
+#include <ctime>
+
+#include <wrench/services/storage/xrootd/XRootD.h>
+
+#include <wrench/services/storage/xrootd/Node.h>
+
+using namespace std;
+const string NETWORK_SPEED="1GBps";
+const string CPU_SPEED="1Gf";
+const string NETWORK_LATENCY="10ms";
+
+namespace sg4 = simgrid::s4u;
+struct Tree{
+	vector<std::shared_ptr<Tree>> children;
+	Tree* parrent;
+	sg4::Host* host;
+};
+void PlatformCreator::operator()() {
+
+        // Create the top-level zone
+        auto zone = sg4::create_full_zone("AS0");
+        auto loopback = zone->create_link("loopback", "1000EBps")->set_latency("0us")-> set_sharing_policy(sg4::Link::SharingPolicy::FATPIPE)->seal();
+        auto fatpipe = zone->create_link("fatpipe", "300MBps")-> set_sharing_policy(sg4::Link::SharingPolicy::FATPIPE)->set_latency("100ms")->seal();
+                // Create a ComputeHost
+        auto user = zone->create_host("user", "1Mf");
+        user->set_core_count(10);
+        user->set_property("ram", "16GB");
+		std::srand(std::time(nullptr));
+		//create graph
+		Tree tRoot;
+		vector<std::shared_ptr<Tree>> prevLevel=vector<std::shared_ptr<Tree>>(leafs);
+		for(int i=0;i<leafs;i++){
+			prevLevel[i]=(std::make_shared<Tree>());
+		}
+		while(true){
+			unsigned int level=round(prevLevel.size()/(64*density));
+			if(level*64<prevLevel.size()){
+				level=(prevLevel.size()/64)+1;//I dont think this will ever happen, but it might due to rounding weirdness, this will just insure that there are always enough nodes
+			}
+			if(level<=1){
+				for(unsigned int i=0;i<prevLevel.size();i++){
+					prevLevel[i]->parrent=&tRoot;
+					tRoot.children.push_back(prevLevel[i]);
+				}
+				break;
+			}
+			vector<std::shared_ptr<Tree>> parrents=vector<std::shared_ptr<Tree>>(level);
+			for(unsigned int i=0;i<level;i++){
+				parrents[i]=(std::make_shared<Tree>());
+			}
+			int unallocated=0;
+			for(unsigned int i=0;i<prevLevel.size();i++){
+				int n=std::rand()/((RAND_MAX + 1u)/level);
+				if(parrents[n]->children.size()<64){
+					prevLevel[i]->parrent=parrents[n].get();
+					parrents[n]->children.push_back(prevLevel[i]);
+				}else{
+					unallocated++;
+				}
+				
+			}
+			int j=0;
+			for(unsigned int i=0;i<prevLevel.size()&&unallocated>0;i++){
+				if(!prevLevel[i]->parrent){
+					while(parrents[j]->children.size()>=64){
+						j=(j+1)%parrents.size();
+					prevLevel[i]->parrent=parrents[j].get();
+					parrents[j]->children.push_back(prevLevel[i]);
+					}
+				}
+			}
+
+			prevLevel=parrents;
+			
+		}
+		
+		
+		auto current=&tRoot;
+		
+		 tRoot.host = zone->create_host("root", "1Mf");
+					tRoot.host->set_core_count(10);
+					tRoot.host->set_property("ram", "16GB");
+
+					//add loopback
+					{
+					sg4::LinkInRoute network_link_in_route{loopback};
+            zone->add_route(tRoot.host->get_netpoint(),
+                            tRoot.host->get_netpoint(),
+                            nullptr,
+                            nullptr,
[…]
