+++ examples/action_api/XRootD/basic-test/src/Controller.cpp
@@ -1,0 +1,205 @@
+
+/**
+ * Copyright (c) 2017-2021. The WRENCH Team.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ */
+
+/**
+ ** An execution controller to execute a workflow
+ **/
+
+#define GFLOP (1000.0 * 1000.0 * 1000.0)
+#define MBYTE (1000.0 * 1000.0)
+#define GBYTE (1000.0 * 1000.0 * 1000.0)
+
+#include <iostream>
+#include <wrench/services/storage/xrootd/Node.h>
+#include "../include/Controller.h"
+std::string padLong(long l){
+	
+	if(l<10){
+		return "0"+std::to_string(l);
+	}else{
+		return std::to_string(l);
+	}
+}
+std::string padDouble(double l){
+	
+	if(l<10){
+		return "0"+std::to_string(l);
+	}else{
+		return std::to_string(l);
+	}
+}
+std::string formatDate(double time){
+	if(time<0){
+		return "Not Started";
+	}
+	long seconds=(long)time;
+	double ms=time-seconds;
+	long minutes=seconds/60;
+	seconds%=60;
+	long hours=minutes/60;
+	minutes%=60;
+	int days=hours/24;
+	hours%=24;
+	
+	return std::to_string(days)+"-"+padLong(hours)+':'+padLong(minutes)+':'+padDouble(seconds+ms);
+}
+WRENCH_LOG_CATEGORY(controller, "Log category for Controller");
+
+namespace wrench {
+
+    /**
+     * @brief Constructor
+     *
+     * @param bare_metal_compute_service: a set of compute services available to run actions
+     * @param storage_services: a set of storage services available to store files
+     * @param hostname: the name of the host on which to start the WMS
+     */
+    Controller::Controller(std::shared_ptr<BareMetalComputeService> bare_metal_compute_service,
+                           const std::shared_ptr<XRootD::Node> &root,
+						   XRootD::XRootD *xrootdManager,
+                           const std::string &hostname) : ExecutionController(hostname, "controller"),
+                                                          bare_metal_compute_service(bare_metal_compute_service), root(root),xrootdManager(xrootdManager) {}
+
+    /**
+     * @brief main method of the Controller
+     *
+     * @return 0 on completion
+     *
+     * @throw std::runtime_error
+     */
+    int Controller::main() {
+
+        /* Set the logging output to GREEN */
+        TerminalOutput::setThisProcessLoggingColor(TerminalOutput::COLOR_GREEN);
+        WRENCH_INFO("Controller starting");
+
+        /* Create a files */
+        std::vector<std::shared_ptr<DataFile>> files ={ 
+			wrench::Simulation::addFile("file1", 1 ),
+			wrench::Simulation::addFile("file2", 1 ),
+			wrench::Simulation::addFile("file3", 1 ),
+			wrench::Simulation::addFile("file4", 1 ),
+			wrench::Simulation::addFile("file5", 1 ),
+			wrench::Simulation::addFile("file6", 1 ),
+			wrench::Simulation::addFile("file7", 1),
+			wrench::Simulation::addFile("file8", 1 ),
+			wrench::Simulation::addFile("file9", 1 ),
+			wrench::Simulation::addFile("file10", 1 ),
+			wrench::Simulation::addFile("file11", 1),
+			wrench::Simulation::addFile("file12", 1 ),
+			wrench::Simulation::addFile("file13", 1 ),
+			wrench::Simulation::addFile("file14", 1 ),
+		};
+        
+			
+        xrootdManager->createFile(files[0], root->getChild(0));//leaf 1
[…]

[…]
