+++ examples/action_api/XRootD/basic-test/CMakeModules/FindSimGrid.cmake
@@ -1,145 +1,0 @@
-
-# CMake find module to search for the SimGrid library.
-
-# Copyright (c) 2016-2022. The SimGrid Team.
-#
-# This file is free software; you can redistribute it and/or modify it
-# under the terms of the license (GNU LGPL) which comes with this package.
-
-#
-# USERS OF PROGRAMS USING SIMGRID
-# -------------------------------
-#
-# If cmake does not find this file, add its path to CMAKE_PREFIX_PATH:
-#    CMAKE_PREFIX_PATH="/path/to/FindSimGrid.cmake:$CMAKE_PREFIX_PATH"  cmake .
-#
-# If this file does not find SimGrid, define SimGrid_PATH:
-#    SimGrid_PATH=/path/to/simgrid  cmake .
-
-#
-# DEVELOPERS OF PROGRAMS USING SIMGRID
-# ------------------------------------
-#
-#  1. Include this file in your own CMakeLists.txt (before defining any target)
-#     Either by copying it in your tree, or (recommended) by using the
-#     version automatically installed by SimGrid.
-#
-#  2. Afterward, if you have CMake >= 2.8.12, this will define a
-#     target called 'SimGrid::Simgrid'. Use it as:
-#       target_link_libraries(your-simulator SimGrid::SimGrid)
-#
-#    With older CMake (< 2.8.12), it simply defines several variables:
-#       SimGrid_INCLUDE_DIR - the SimGrid include directories
-#       SimGrid_LIBRARY - link your simulator against it to use SimGrid
-#    Use as:
-#      include_directories("${SimGrid_INCLUDE_DIR}" SYSTEM)
-#      target_link_libraries(your-simulator ${SimGrid_LIBRARY})
-#
-#  In both cases, it also define a SimGrid_VERSION macro, that you
-#    can use to deal with API evolutions as follows:
-#
-#    #if SimGrid_VERSION < 31800
-#      (code to use if the installed version is lower than v3.18)
-#    #elif SimGrid_VERSION < 31900
-#      (code to use if we are using SimGrid v3.18.x)
-#    #else
-#      (code to use with SimGrid v3.19+)
-#    #endif
-#
-#  Since SimGrid header files require C++14, so we set CMAKE_CXX_STANDARD to 14.
-#    Change this variable in your own file if you need a later standard.
-
-#
-# IMPROVING THIS FILE
-# -------------------
-#  - Use automatic SimGridConfig.cmake creation via export/install(EXPORT in main CMakeLists.txt:
-#    https://cliutils.gitlab.io/modern-cmake/chapters/exporting.html
-#    https://cmake.org/Wiki/CMake/Tutorials/How_to_create_a_ProjectConfig.cmake_file
-#    https://github.com/boostcon/cppnow_presentations_2017/blob/master/05-19-2017_friday/effective_cmake__daniel_pfeifer__cppnow_05-19-2017.pdf
-
-cmake_minimum_required(VERSION 2.8.12)
-
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-find_path(SimGrid_INCLUDE_DIR
-  NAMES simgrid/config.h
-  NAMES simgrid/version.h
[…]
