### 2022-09-01 70959f1
removed stress test from examples
