### 2022-09-01 308f3f7
removed stress test from examples
