### 2022-09-01 c2db398
removed stress test from examples
