### 2022-09-01 48e05f5
Some code changes / cleanup / formatting
