+++ examples/run_all_examples.sh.in
@@ -64,7 +64,7 @@ ${INSTALL_DIR}/workflow_api/real-workflow-example/wrench-example-real-workflow $
 echo "${bold}RUNNING: ${INSTALL_DIR}/workflow_api/condor-grid-example/wrench-example-condor-grid-universe${normal}"
 ${INSTALL_DIR}/workflow_api/condor-grid-example/wrench-example-condor-grid-universe ${INSTALL_DIR}/workflow_api/condor-grid-example/condor-grid-hosts.xml
 
-echo "${bold}RUNNING: ${INSTALL_DIR}/action_api/XRootD/basic-test/wrench-example-xrootd-basic platform.xml false${normal}"
-${INSTALL_DIR}/action_api/XRootD/basic-test/wrench-example-xrootd-basic ${INSTALL_DIR}/action_api/XRootD/basic-test/platform.xml
+echo "${bold}RUNNING: ${INSTALL_DIR}/action_api/XRootD/wrench-example-xrootd-basic platform.xml false${normal}"
+${INSTALL_DIR}/action_api/XRootD/wrench-example-xrootd-basic ${INSTALL_DIR}/action_api/XRootD/basic-test/platform.xml
 
 
