### 2022-09-01 2ffdc13
Some code changes / cleanup / formatting
