+++ conf/cmake/Examples.cmake
@@ -22,7 +22,7 @@ add_subdirectory(examples/action_api/multi-action-multi-job EXCLUDE_FROM_ALL)
 add_subdirectory(examples/action_api/job-action-failure EXCLUDE_FROM_ALL)
 add_subdirectory(examples/action_api/super-custom-action EXCLUDE_FROM_ALL)
 
-add_subdirectory(examples/action_api/XRootD/basic-test/ EXCLUDE_FROM_ALL)
+add_subdirectory(examples/action_api/XRootD/ EXCLUDE_FROM_ALL)
 
 add_custom_command(TARGET examples
     COMMAND cat ${CMAKE_CURRENT_SOURCE_DIR}/examples/run_all_examples.sh.in | sed "s~TO_FILL_IN~${CMAKE_BINARY_DIR}/examples/~g" > ${CMAKE_BINARY_DIR}/examples/run_all_examples.sh
