+++ include/wrench/services/storage/xrootd/XRootDDeployment.h
@@ -15,8 +15,8 @@
  * (at your option) any later version.
  */
 
-#ifndef WRENCH_XROOTD_H
-#define WRENCH_XROOTD_H
+#ifndef WRENCH_XROOTDDEPLOYMENT_H
+#define WRENCH_XROOTDDEPLOYMENT_H
 #include "wrench/services/Service.h"
 #include <vector>
 #include <unordered_map>
@@ -31,11 +31,8 @@ namespace wrench {
      * @brief A Meta manager for an XRootD data Federation.  This tracks all nodes and files within the system.
      */
     namespace XRootD{
-        //class StorageServer;
-        //class Supervisor;
-//        class Node;
 
-        class XRootD{
+        class XRootDDeployment {
         public:
 
             /**
@@ -44,18 +41,19 @@ namespace wrench {
              * @param property_values: The property values that should be used to overwrite the defaults of all Nodes (defaults to none) (unless otherwise specified)
              * @param messagepayload_values: The message paylaod values that should be used to overwrite the defaults of all Nodes (defaults to none) (unless otherwise specified)
              */
-            XRootD(std::shared_ptr<Simulation>  simulation,WRENCH_PROPERTY_COLLECTION_TYPE property_values={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_values={}):property_values(property_values),messagepayload_values(messagepayload_values),simulation(simulation){}
-            std::shared_ptr<Node> createStorageServer(const std::string& hostname, const std::string& mount_point, WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list,WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
-            std::shared_ptr<Node> createSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
-            std::shared_ptr<Node> createStorageSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE storage_property_list, WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE storage_messagepayload_list,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
-            ~XRootD(){}
+            XRootDDeployment(std::shared_ptr<Simulation>  simulation, WRENCH_PROPERTY_COLLECTION_TYPE property_values={}, WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE messagepayload_values={}): property_values(property_values), messagepayload_values(messagepayload_values), simulation(simulation){}
+
+            std::shared_ptr<Node> createRootSupervisor(const std::string& hostname,WRENCH_PROPERTY_COLLECTION_TYPE node_property_list={},WRENCH_MESSAGE_PAYLOADCOLLECTION_TYPE node_messagepayload_list={});
+            std::shared_ptr<Node> getRootSupervisor();
+
+            ~XRootDDeployment(){}
             /***********************/
             /** \cond DEVELOPER    */
             /***********************/
             /** @brief The max number of hops a search message can take.  Used to prevent infinite loops in a poorly constructed XRootD tree. */
             int defaultTimeToLive=1024;//how long trivial search message can wander for;
             
[…]

[…]
