+++ examples/action_api/XRootD/Controller.cpp
@@ -9,7 +9,8 @@
  */
 
 /**
- ** An execution controller to execute a workflow
+ ** An execution controller to execute a multi-action job, where the actions
+ ** read files from and XRootD deployment
  **/
 
 #define GFLOP (1000.0 * 1000.0 * 1000.0)
@@ -34,17 +35,17 @@ std::string padDouble(double l){
 }
 
 std::string formatDate(double time){
-	if(time<0){
+	if(time < 0){
 		return "Not Started";
 	}
-	long seconds=(long)time;
+	long seconds = (long)time;
 	double ms = time - (double)seconds;
 	long minutes = seconds / 60;
 	seconds %= 60;
-	long hours=minutes/60;
-	minutes%=60;
-	long days=hours/24;
-	hours%=24;
+	long hours = minutes / 60;
+	minutes %= 60;
+	long days = hours / 24;
+	hours %= 24;
 	
 	return std::to_string(days)+"-"+padLong(hours)+':'+padLong(minutes)+':'+padDouble((double)seconds+ms);
 }
@@ -56,15 +57,16 @@ namespace wrench {
     /**
      * @brief Constructor
      *
-     * @param bare_metal_compute_service: a set of compute services available to run actions
-     * @param storage_services: a set of storage services available to store files
-     * @param hostname: the name of the host on which to start the WMS
+     * @param bare_metal_compute_service: a bare-metal compute services available to run actions
+     * @param root: The root node of an XRootD deployment
+     * @param xrootd_manager: The XRootD manager
+     * @param hostname: the name of the host on which to start the Controller
      */
     Controller::Controller(const std::shared_ptr<BareMetalComputeService> &bare_metal_compute_service,
-                           const std::shared_ptr<XRootD::Node> &root,
-						   XRootD::XRootD *xrootdManager,
+						   XRootD::XRootDDeployment *xrootd_deployment,
                            const std::string &hostname) : ExecutionController(hostname, "controller"),
-                                                          bare_metal_compute_service(bare_metal_compute_service), root(root),xrootdManager(xrootdManager) {}
+                                                          bare_metal_compute_service(bare_metal_compute_service),
+                                                          xrootd_deployment(xrootd_deployment) {}
 
     /**
      * @brief main method of the Controller
@@ -99,23 +101,28 @@ namespace wrench {
 			wrench::Simulation::addFile("file14", 1 ),
 		};
         
-			
-        xrootdManager->createFile(files[0], root->getChild(0));//leaf 1
-		xrootdManager->createFile(files[1], root->getChild(1));//leaf 2
-		xrootdManager->createFile(files[2], root->getChild(2)->getChild(0));//leaf 3
-		xrootdManager->createFile(files[3], root->getChild(2)->getChild(1));//leaf 4
+        auto root = this->xrootd_deployment->getRootSupervisor();
+
+
+        root->getChild(0)->createFile(files[0]);
+        root->getChild(1)->createFile(files[1]);
+        root->getChild(2)->getChild(0)->createFile(files[2]);
+        root->getChild(2)->getChild(1)->createFile(files[3]);
 		root->getChild(2)->getChild(2)->getChild(0)->createFile(files[4]);//leaf 5
-		root->getChild(2)->getChild(2)->getChild(1)->createFile(files[5]);//leaf 6
-		root->getChild(2)->getChild(2)->getChild(2)->getChild(0)->createFile(files[6]);//leaf 7
[…]
