+++ include/wrench.h
@@ -17,7 +17,7 @@
 #include "wrench/services/compute/bare_metal/BareMetalComputeServiceProperty.h"
 #include "wrench/services/storage/simple/SimpleStorageService.h"
 #include "wrench/services/storage/simple/SimpleStorageServiceProperty.h"
-#include "wrench/services/storage/xrootd/XRootD.h"
+#include "wrench/services/storage/xrootd/XRootDDeployment.h"
 #include "wrench/services/storage/xrootd/XRootDProperty.h"
 #include "wrench/services/file_registry/FileRegistryService.h"
 #include "wrench/services/file_registry/FileRegistryServiceProperty.h"
