+++ CMakeLists.txt
@@ -186,7 +186,7 @@ set(HEADER_FILES
         include/wrench/execution_controller/ExecutionController.h
         include/wrench/execution_controller/ExecutionControllerMessage.h
         include/wrench/data_file/DataFile.h
-        include/wrench/services/storage/xrootd/XRootD.h
+        include/wrench/services/storage/xrootd/XRootDDeployment.h
         include/wrench/services/storage/xrootd/Node.h
         include/wrench/services/storage/xrootd/XRootDMessagePayload.h
         include/wrench/services/storage/xrootd/XRootDMessage.h
@@ -354,7 +354,7 @@ set(SOURCE_FILES
         src/wrench/execution_controller/ExecutionController.cpp
         src/wrench/execution_controller/ExecutionControllerMessage.cpp
         src/wrench/data_file/DataFile.cpp
-        src/wrench/services/storage/xrootd/XRootD.cpp
+        src/wrench/services/storage/xrootd/XRootDDeployment.cpp
         src/wrench/services/storage/xrootd/Node.cpp
         src/wrench/services/storage/xrootd/SearchStack.cpp
         src/wrench/services/storage/xrootd/Cache.cpp

[…]
