+++ test/services/storage_services/XRootDStorageService/XRootDStorageServiceFullFunctionalTest.cpp
@@ -1,0 +1,299 @@
+/**
+ * Copyright (c) 2017. The WRENCH Team.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ */
+
+
+#include <gtest/gtest.h>
+
+#include <wrench-dev.h>
+#include "../../../include/TestWithFork.h"
+#include "../../../include/UniqueTmpPathPrefix.h"
+
+#define GFLOP (1000.0 * 1000.0 * 1000.0)
+#define MBYTE (1000.0 * 1000.0)
+#define GBYTE (1000.0 * 1000.0 * 1000.0)
+
+WRENCH_LOG_CATEGORY(xrootd_storage_service_full_functional_test, "Log category for XRootDServiceFullFunctionalTest");
+
+
+/*
+ * Helper function for pretty-printed output
+ */
+std::string padLong(long l){
+    return (l < 10 ? "0"+std::to_string(l) : std::to_string(l));
+}
+
+std::string padDouble(double l){
+    return (l < 10 ? "0"+std::to_string(l) : std::to_string(l));
+}
+
+std::string formatDate(double time){
+    if(time < 0){
+        return "Not Started";
+    }
+    long seconds = (long)time;
+    double ms = time - (double)seconds;
+    long minutes = seconds / 60;
+    seconds %= 60;
+    long hours = minutes / 60;
+    minutes %= 60;
+    long days = hours / 24;
+    hours %= 24;
+
+    return std::to_string(days)+"-"+padLong(hours)+':'+padLong(minutes)+':'+padDouble((double)seconds+ms);
+}
+
+class XRootDServiceFunctionalTest : public ::testing::Test {
+
+public:
+
+    void do_FullFunctionality_test();
+
+    std::shared_ptr<wrench::XRootD::Node> root_supervisor;
+
+protected:
+
+    XRootDServiceFunctionalTest() {
+    }
+
+};
+
+
+/**********************************************************************/
+/**  FULL FUNCTIONALITY SIMULATION TEST                             **/
+/**********************************************************************/
+
+class XRootDServiceFullFunctionalityTestExecutionController : public wrench::ExecutionController {
+
+public:
+    XRootDServiceFullFunctionalityTestExecutionController(XRootDServiceFunctionalTest *test,
+                                                          wrench::XRootD::XRootDDeployment *xrootd_deployment,
+                                                          std::shared_ptr<wrench::BareMetalComputeService> bare_metal_compute_service,
+                                                          std::string hostname) :
+            wrench::ExecutionController(hostname, "test"), test(test),
+            xrootd_deployment(xrootd_deployment),
+            bare_metal_compute_service(bare_metal_compute_service) {}
+
+private:
+    XRootDServiceFunctionalTest *test;
+    wrench::XRootD::XRootDDeployment *xrootd_deployment;
+    std::shared_ptr<wrench::BareMetalComputeService> bare_metal_compute_service;
+
+    int main() override {
+        /* Set the logging output to GREEN */
+        wrench::TerminalOutput::setThisProcessLoggingColor(wrench::TerminalOutput::COLOR_GREEN);
+        WRENCH_INFO("Controller starting");
+
+        /* Add a bunch of 1-byte files to the simulation, which will
+         * then be stored at storage servers in the XRootD tree
+         */
+        std::vector<std::shared_ptr<wrench::DataFile>> files ={
+                wrench::Simulation::addFile("file01", 1 ),
+                wrench::Simulation::addFile("file02", 1 ),
+                wrench::Simulation::addFile("file03", 1 ),
+                wrench::Simulation::addFile("file04", 1 ),
[…]
