### 2022-09-02 c712705
Added a unit test
