+++ examples/action_api/XRootD/Controller.cpp
@@ -27,27 +27,27 @@
  * Helper function for pretty-printed output
  */
 std::string padLong(long l){
-	return (l < 10 ? "0"+std::to_string(l) : std::to_string(l));
+    return (l < 10 ? "0"+std::to_string(l) : std::to_string(l));
 }
 
 std::string padDouble(double l){
-	return (l < 10 ? "0"+std::to_string(l) : std::to_string(l));
+    return (l < 10 ? "0"+std::to_string(l) : std::to_string(l));
 }
 
 std::string formatDate(double time){
-	if(time < 0){
-		return "Not Started";
-	}
-	long seconds = (long)time;
-	double ms = time - (double)seconds;
-	long minutes = seconds / 60;
-	seconds %= 60;
-	long hours = minutes / 60;
-	minutes %= 60;
-	long days = hours / 24;
-	hours %= 24;
-	
-	return std::to_string(days)+"-"+padLong(hours)+':'+padLong(minutes)+':'+padDouble((double)seconds+ms);
+    if(time < 0){
+        return "Not Started";
+    }
+    long seconds = (long)time;
+    double ms = time - (double)seconds;
+    long minutes = seconds / 60;
+    seconds %= 60;
+    long hours = minutes / 60;
+    minutes %= 60;
+    long days = hours / 24;
+    hours %= 24;
+
+    return std::to_string(days)+"-"+padLong(hours)+':'+padLong(minutes)+':'+padDouble((double)seconds+ms);
 }
 
 WRENCH_LOG_CATEGORY(controller, "Log category for Controller");
@@ -63,7 +63,7 @@ namespace wrench {
      * @param hostname: the name of the host on which to start the Controller
      */
     Controller::Controller(const std::shared_ptr<BareMetalComputeService> &bare_metal_compute_service,
-						   XRootD::XRootDDeployment *xrootd_deployment,
+                           XRootD::XRootDDeployment *xrootd_deployment,
                            const std::string &hostname) : ExecutionController(hostname, "controller"),
                                                           bare_metal_compute_service(bare_metal_compute_service),
                                                           xrootd_deployment(xrootd_deployment) {}
@@ -84,118 +84,84 @@ namespace wrench {
         /* Add a bunch of 1-byte files to the simulation, which will
          * then be stored at storage servers in the XRootD tree
          */
-        std::vector<std::shared_ptr<DataFile>> files ={ 
-			wrench::Simulation::addFile("file01", 1 ),
-			wrench::Simulation::addFile("file02", 1 ),
-			wrench::Simulation::addFile("file03", 1 ),
-			wrench::Simulation::addFile("file04", 1 ),
-			wrench::Simulation::addFile("file05", 1 ),
-			wrench::Simulation::addFile("file06", 1 ),
-			wrench::Simulation::addFile("file07", 1),
-			wrench::Simulation::addFile("file08", 1 ),
-			wrench::Simulation::addFile("file09", 1 ),
-			wrench::Simulation::addFile("file10", 1 ),
-			wrench::Simulation::addFile("file11", 1),
-			wrench::Simulation::addFile("file12", 1 ),
-			wrench::Simulation::addFile("file13", 1 ),
-			wrench::Simulation::addFile("file14", 1 ),
-		};
-        
+        auto some_file = wrench::Simulation::addFile("some_file", 12 * MBYTE);
+
+        /* This is the XRootD tree that was constructed before the simulation was launched.
+         * We will create replicas of the "some_file" file at nodes leaf4, leaf8, leaf9, which
+         * we marked with a '*' below.
+
+            root
+         /   |   \
+    leaf1  leaf2  super1
+                 /   |   \
+            leaf3  leaf4*  super2
+                         /   |   \
+                    leaf5  leaf6  super3
+                                 /   |   \
[…]
