+++ examples/action_api/XRootD/Controller.h
@@ -32,6 +32,7 @@ namespace wrench {
     protected:
         // Overridden method
         void processEventCompoundJobCompletion(std::shared_ptr<CompoundJobCompletedEvent>) override;
+        void processEventCompoundJobFailure(std::shared_ptr<CompoundJobFailedEvent> event) override;
 
     private:
         // main() method of the Controller
