### 2022-09-03 34eac4e
Committing clang-format changes
