+++ src/wrench/services/storage/xrootd/Node.cpp
@@ -36,12 +36,12 @@ namespace wrench {
             //    internalStorage->start(internalStorage,true,true);
             //}
             std::string message =
-                    "XRootD Node " + this->getName() + "  starting on host " + this->getHostname();WRENCH_INFO("%s",
-                                                                                                               message.c_str());
-            reduced= getPropertyValueAsBoolean(Property::REDUCED_SIMULATION);
+                    "XRootD Node " + this->getName() + "  starting on host " + this->getHostname();
+            WRENCH_INFO("%s",
+                        message.c_str());
+            reduced = getPropertyValueAsBoolean(Property::REDUCED_SIMULATION);
             /** Main loop **/
             while (this->processNextMessage()) {
-
             }
 
             WRENCH_INFO("XRootD Node %s on host %s cleanly terminating!",
@@ -57,7 +57,7 @@ namespace wrench {
          * @return the child id of the new node
          * @throws runtime_error if node already has 64 children.
          */
-         //should this throw an exception instead?
+        //should this throw an exception instead?
         std::shared_ptr<Node> Node::addChild(std::shared_ptr<Node> child) {
             child->supervisor = this;
             if (children.size() < 64) {
@@ -66,49 +66,49 @@ namespace wrench {
             }
             throw std::runtime_error("Supervisor running on " + hostname + " already has 64 children");
         }
-//            /**
-//             * @brief Get mount point of underlying storage service
-//             * @return the (sole) mount point of the service if there is one, and null otherwise
-//             */
-//            std::string Node::getMountPoint(){
-//                if(internalStorage){
-//                    return internalStorage->getMountPoint();
-//                }
-//                return "";
-//            }
-//            /**
-//             * @brief Get the set of mount points of the underlying storage service
-//             * @return the set of mount points of the service if there is one, and empty set otherwise
-//             */
-//            std::set<std::string> Node::getMountPoints() {
-//                if(internalStorage){
-//                    return internalStorage->getMountPoints();
-//                }
-//                return std::set<std::string>();
-//            }
-//            /**
-//             * @brief Check whether the underlying storage service has multiple mount points
-//             * @return whether the service has multiple mount points, if there is no storage service, returns false
-//             */
-//            bool Node::hasMultipleMountPoints(){
-//                if(internalStorage){
-//                    return internalStorage->hasMultipleMountPoints();
-//                }
-//                return false;
-//            }
-//            /**
-//            * @brief Check whether the underlying storage service has a particular mount point
-//            * @param mp: a mount point
-//            *
-//            * @return whether the underlying service has that mount point, if there is no storage service, returns false
-//            */
-//            bool Node::hasMountPoint(const std::string &mp){
-//                if(internalStorage){
-//                    return internalStorage->hasMultipleMountPoints();
-//                }
-//                return false;
-//
-//            }
+        //            /**
+        //             * @brief Get mount point of underlying storage service
+        //             * @return the (sole) mount point of the service if there is one, and null otherwise
+        //             */
+        //            std::string Node::getMountPoint(){
+        //                if(internalStorage){
+        //                    return internalStorage->getMountPoint();
+        //                }
+        //                return "";
+        //            }
+        //            /**
+        //             * @brief Get the set of mount points of the underlying storage service
+        //             * @return the set of mount points of the service if there is one, and empty set otherwise
+        //             */
+        //            std::set<std::string> Node::getMountPoints() {
+        //                if(internalStorage){
+        //                    return internalStorage->getMountPoints();
+        //                }
+        //                return std::set<std::string>();
+        //            }
+        //            /**
+        //             * @brief Check whether the underlying storage service has multiple mount points
+        //             * @return whether the service has multiple mount points, if there is no storage service, returns false
+        //             */
[…]
