+++ include/wrench/job/CompoundJob.h
@@ -75,27 +75,26 @@ namespace wrench {
 
         std::shared_ptr<FileReadAction> addFileReadAction(const std::string &name,
                                                           const std::shared_ptr<DataFile> &file,
-                                                          const std::shared_ptr<StorageService>& storageService);
+                                                          const std::shared_ptr<StorageService> &storageService);
 
         std::shared_ptr<FileReadAction> addFileReadAction(const std::string &name,
                                                           const std::shared_ptr<DataFile> &file,
-                                                          const std::shared_ptr<StorageService>& storageService,
+                                                          const std::shared_ptr<StorageService> &storageService,
                                                           const double num_bytes_to_read);
 
 
         std::shared_ptr<FileWriteAction> addFileWriteAction(const std::string &name,
                                                             const std::shared_ptr<DataFile> &file,
-                                                            const std::shared_ptr<StorageService>& storageService);
+                                                            const std::shared_ptr<StorageService> &storageService);
 
         std::shared_ptr<FileCopyAction> addFileCopyAction(const std::string &name,
                                                           std::shared_ptr<DataFile> file,
-                                                          const std::shared_ptr<StorageService>& src_storageService,
-                                                          const std::shared_ptr<StorageService>& dest_storageService);
+                                                          const std::shared_ptr<StorageService> &src_storageService,
+                                                          const std::shared_ptr<StorageService> &dest_storageService);
 
         std::shared_ptr<FileDeleteAction> addFileDeleteAction(const std::string &name,
                                                               std::shared_ptr<DataFile> file,
-                                                              const std::shared_ptr<StorageService>& storageService);
-
+                                                              const std::shared_ptr<StorageService> &storageService);
 
 
         std::shared_ptr<FileReadAction> addFileReadAction(const std::string &name,

[…]
