### 2022-09-03 3101714
Committing clang-format changes
