+++ src/wrench/simulation/Simulation.cpp
@@ -756,7 +756,7 @@ namespace wrench {
     void Simulation::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<FileLocation> &location) {
         //location->getStorageService()->stageFile(file, location->getMountPoint(),
         //                                         location->getAbsolutePathAtMountPoint());
-        location->getStorageService()->createFile(file,location);
+        location->getStorageService()->createFile(file, location);
     }
     /**
      * @brief Store a file on a particular file server ex-nihilo. Doesn't notify a file registry service and will do nothing (and won't complain) if the file already exists
@@ -766,8 +766,7 @@ namespace wrench {
      *
      * @throw std::invalid_argument
      */
-    [[deprecated("Replaced by StorageService::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<FileLocation> &location), do not use if using XRootD or other distributed file system")]]
-    void Simulation::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<StorageService> &server) {
+    [[deprecated("Replaced by StorageService::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<FileLocation> &location), do not use if using XRootD or other distributed file system")]] void Simulation::createFile(const std::shared_ptr<DataFile> &file, const std::shared_ptr<StorageService> &server) {
         server->createFile(file);
     }
 
