+++ include/wrench/services/storage/xrootd/XRootDMessage.h
@@ -28,7 +28,7 @@
 #include "wrench/services/storage/StorageServiceMessage.h"
 
 namespace wrench {
-    namespace XRootD{
+    namespace XRootD {
 
         /***********************/
         /** \cond INTERNAL     */
@@ -49,22 +49,22 @@ namespace wrench {
             ContinueSearchMessage(simgrid::s4u::Mailbox *answer_mailbox,
                                   std::shared_ptr<StorageServiceFileReadRequestMessage> original,
                                   std::shared_ptr<DataFile> file,
-                                  Node* node,
+                                  Node *node,
                                   double payload,
                                   std::shared_ptr<bool> answered,
                                   int timeToLive);
-            ContinueSearchMessage(ContinueSearchMessage* toCopy);
+            ContinueSearchMessage(ContinueSearchMessage *toCopy);
             /** @brief Mailbox to which the FINAL answer message should be sent */
             simgrid::s4u::Mailbox *answer_mailbox;
 
-            /** @brief The original file read request that kicked off the search (if null this was a lookup request)*/\
+            /** @brief The original file read request that kicked off the search (if null this was a lookup request)*/
             std::shared_ptr<StorageServiceFileReadRequestMessage> original;
 
             /** @brief The file being searched for */
             std::shared_ptr<DataFile> file;
 
             /** The node that originally received the FileLookupRequest or FileReadRequest */
-            Node* node;
+            Node *node;
             /** @brief Whether or not the calling client has been answered yet.  Used to prevent answer_mailbox spamming for multiple file hits */
             std::shared_ptr<bool> answered;
             /** How many more hops this message can live for, to prevent messages living forever in improper configurations with loops.*/
@@ -77,20 +77,20 @@ namespace wrench {
 
         class UpdateCacheMessage : public Message {
         public:
-            UpdateCacheMessage(simgrid::s4u::Mailbox *answer_mailbox,std::shared_ptr<StorageServiceFileReadRequestMessage> original,Node* node,std::shared_ptr<DataFile> file,  std::set<std::shared_ptr<FileLocation>> locations,
+            UpdateCacheMessage(simgrid::s4u::Mailbox *answer_mailbox, std::shared_ptr<StorageServiceFileReadRequestMessage> original, Node *node, std::shared_ptr<DataFile> file, std::set<std::shared_ptr<FileLocation>> locations,
                                double payload, std::shared_ptr<bool> answered);
-            UpdateCacheMessage(UpdateCacheMessage& other);
-            UpdateCacheMessage(UpdateCacheMessage* other);
+            UpdateCacheMessage(UpdateCacheMessage &other);
+            UpdateCacheMessage(UpdateCacheMessage *other);
             /** @brief Mailbox to which the FINAL answer message should be sent */
             simgrid::s4u::Mailbox *answer_mailbox;
-            /** @brief The original file read request that kicked off the search (if null this was a lookup request)*/\
+            /** @brief The original file read request that kicked off the search (if null this was a lookup request)*/
             std::shared_ptr<StorageServiceFileReadRequestMessage> original;
             /** @brief The file found */
             std::shared_ptr<DataFile> file;
             /** @brief the locations to cache */
             set<std::shared_ptr<FileLocation>> locations;
             /** @brief The highest node in the tree to return to when caching (should be the node the original message was sent) */
-            Node* node;
+            Node *node;
             /** @brief Whether or not the calling client has been answered yet.  Used to prevent answer_mailbox spamming for multiple file hits */
             std::shared_ptr<bool> answered;
         };
@@ -100,53 +100,52 @@ namespace wrench {
          */
[…]
