### 2022-09-19 #4772 @editorialbot
[REVIEW]: WRENCH 2.0: Cyberinfrastructure simulation workbench
**Submitting author:** @henricasanova (<a href="http://orcid.org/0000-0001-6310-0365">Henri Casanova</a>)
**Repository:** https://github.com/wrench-project/wrench
**Branch with paper.md** (empty if default branch): joss_2022
**Version:** v2.0
