### 2022-09-21 #4772 @jarrah42
@gkthiruvathukal One of the authors works at the same organization as me, but not in the same part of the organization. I have collaborated with him in the past and have had recent interactions with him. Does that constitute a COI for this review?
