### 2023-02-02 #4772 @yadudoc
@gkthiruvathukal, I went back through my notes and found that I had gotten stuck on installation and this review fell through the cracks. Apologies for the delay on this.

@henricasanova, I'm stuck on the installation of simgrid on my mac, and I suspect this might be an apple silicon issue. It's not an issue on `wrench` so I'm not sure who to approach about this. For reference, this is a full paste of the error I'm getting: https://gist.github.com/yadudoc/6c95964ce0d44565ba466d30a10f20c4.
