### 2023-02-02 #4772 @henricasanova
@yadudoc Let me investigate. I am installing SimGrid on my M1 mac without problems.  I have cmake version 3.24.3 (while you have 3.22.1), but I can't imagine that's a problem.  In terms of the compile I am using the default /usr/bin/clang, which is version 14.0.0.   I see that you're using installed using anaconda, and it's version 13.4.  Perhaps that's the problem? I have never seen that error before.  Would it help if I created a Docker container for you perhaps?
