### 2023-09-14 #4772 @henricasanova
@gkthiruvathukal Apologies, I never saw this notification. In July I was on travel to deal with a pretty dire personal situation, and sadly a lot of things fell through the cracks.I was waiting for an answer regarding my "Would it help if I created a Docker container for you perhaps?" question to yadudoc, since, for some reason, he couldn't build on of WRENCH's dependencies on his new M1 laptop.  I was ready to do that of course. That was in February.  So at this point, WRENCH has had two new releases and this has been going on for too long. So I think it's just saner to just drop this whole affair. thanks.
