### 2023-09-14 #4772 @danielskatz
@henricasanova - On behalf of JOSS, I apologize for the slow/missing interactions. Our communications process between editor, reviewers, and author appears to have broken down here.
